% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R
\name{accuracy}
\alias{accuracy}
\title{Overall Model Accuracy}
\usage{
accuracy(object, formula = "corr", na.rm = FALSE)
}
\arguments{
\item{x}{An object that inherits from class \code{"rpart"}.}
}
\description{
\code{accuracy} returns a single number between zero and one describing the
overall accuracy of the decision tree.
}
\keyword{internal}

