% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_at.R
\name{add_at}
\alias{add_at}
\alias{rm_at}
\title{Add or remove \code{_at} from string}
\usage{
add_at(x)

rm_at(x)
}
\arguments{
\item{x}{A character vector of feature name}
}
\value{
\code{add_at} and \code{rm_at} both return a character of the same
  length as \code{x} with and without \code{"_at"}, respectively.
}
\description{
Adds (\code{add_at}) or removes (\code{add_at}) the string \code{"_at"} 
suffix from a character vector. The \code{"_at"} is often added by various  
bioconductor microarray preprocessing packages.
}
\examples{
add_at(c("hello", "world", "with_at"))
rm_at(c("hello_at", "world_at", "nc"))
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
