% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{add_cov_to_formula}
\alias{add_cov_to_formula}
\title{Add a Covariate to a Formula}
\usage{
add_cov_to_formula(covs, formula)
}
\arguments{
\item{covs}{should be a list of variable names}

\item{formula}{which formula to add covariates to}
}
\value{
formula
}
\description{
\code{add_cov_to_formula} adds some covariates to a formula;
  covs should be a list of variable names
}
\examples{
ff <- y ~ x
add_cov_to_formula(list("w", "z"), ff)

ff <- ~x
add_cov_to_formula("z", ff)

}
