% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{addCovToFormla}
\alias{addCovToFormla}
\title{addCovToFormla}
\usage{
addCovToFormla(covs, formla)
}
\arguments{
\item{covs}{should be a list of variable names}
}
\description{
Legacy version of `add_cov_to_formula`, please
 use that function instead.  This function will eventually be
 deleted.
}
\keyword{internal}
