% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hybrid.R
\name{assignNode}
\alias{assignNode}
\title{Terminal Node Assignment}
\usage{
assignNode(object, newdata, na.action = na.pass, ...)
}
\arguments{
\item{object}{An object that inherits from class \code{rpart}.}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{na.action}{a function which indicates what should happen when the
data contain \code{NA}s.}

\item{...}{Additional optional arguments. At present, no optional arguments
  are used.}
}
\value{
A numeric vector containing the terminal node each observation
  belongs to.
}
\description{
Assign each observation of the original data (or newdata) to a terminal
node. Based on a nice hack from stackoverflow.
}

