% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_pattern_analysis.R
\name{basic_pattern_analysis}
\alias{basic_pattern_analysis}
\title{Basic Pattern Analysis}
\usage{
basic_pattern_analysis(x, unique_only = FALSE, ...)
}
\arguments{
\item{x}{A data frame or character vector.}

\item{unique_only}{Logical indicating wether or not to only show the unique
patterns. Default is \code{FALSE}.}

\item{...}{Additional optional arguments. (Currently ignored).}
}
\description{
Perform a basic pattern analysis
}
\examples{
basic_pattern_analysis(iris)
basic_pattern_analysis(iris, unique_only = TRUE)
}

