% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listBatteryMethods}
\alias{battery}
\alias{battery.rpart}
\alias{listBatteryMethods}
\title{List of battery methods}
\usage{
listBatteryMethods()

battery(object, ...)

\method{battery}{rpart}(object, type = c("priors", "shaving", "target",
  "depth"), newdata, .progress = TRUE, prior.grid, prior.min = 0.05,
  prior.max = 0.95, prior.step = 0.05, ...)
}
\arguments{
\item{object}{An object that inherits from class \code{rpart} or
\code{randomForest}.}

\item{...}{Additional optional arguments to be passed onto \code{rpart} or
           \code{randomForest}.}

\item{battery}{The type of battery to perform. For the entire list of
               battery methods use \code{batteryMethods("rpart")}.}
}
\value{
A list with two components:
\itemize{
  \item \code{rpart}
  \item \code{randomForest}
}

A data frame.
}
\description{
Prints a list of the available battery methods.

Tree-based algorithms are characterized by a substantial number of control
settings. Often the optimal values for many parameters cannot be determined
beforehand and require a trial and error experimental approach. In other
cases, it is desirable to try various settings to study their impact on the
resulting models. CART batteries were designed to automate the most
frequently-occurring modeling situations that require multiple collections
of CART runs.
}
\examples{
## Spam example
library(rpart)
data(spam, package = "kernlab")
spam_cart <- rpart(type ~ ., data = spam)

## Battery priors
table(spam$type)/nrow(spam)  # default priors
battery(spam_cart, type = "priors")
}
\references{
Salford Systems
}

