% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R
\name{battery_target}
\alias{battery_target}
\title{Battery Target}
\usage{
battery_target(X, ...)
}
\arguments{
\item{X}{A data frame of predictor variables.}

\item{...}{Additional optional arguments to be passed on to \code{rpart}.}
}
\description{
Automatic extraction of mutual dependencies among predictors. Here, the word
"dependencies," refers to a potentially nonlinear multivariate relationship.
}
\examples{
data(mushroom, package = "misc")
bt <- battery_target(Mushroom)
bt
plot(bt)
}

