% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{blockBootSample}
\alias{blockBootSample}
\title{blockBootSample}
\usage{
blockBootSample(data, idname)
}
\arguments{
\item{data}{data.frame from which you want to bootstrap}

\item{idname}{column in data which contains an individual identifier}
}
\value{
data.frame bootstrapped from the original dataset; this data.frame
 will contain new ids
}
\description{
make draws of all observations with the same id in a panel
 data context.  This is useful for bootstrapping with panel data.
}
