% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bmisc-package.R
\docType{package}
\name{Bmisc-package}
\alias{Bmisc}
\alias{Bmisc-package}
\alias{bmisc}
\title{My biostatistical miscellaneous functions}
\description{
My frequently used biostatistical miscellaneous R functions and utilities.
Useful for general microarray analysis, gene network analysis, and other
biostatistical applications.
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}

