% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bSpline}
\alias{bSpline}
\title{Plot piecewise B-splines}
\usage{
bSpline(x, y, order = 4, n.evals = 100)
}
\arguments{
\item{x}{A vector giving the x-coordinates of the control points.
Alternatively, \code{x} can be a \code{list} or \code{data.frame} where
the two first elements (or columns) are used.}

\item{y}{An optional vector giving the y-coordinates of the control points.}

\item{order}{the order of the B-splines. Default is cubic (\code{order = 4}).}

\item{n.evals}{integer giving the number of points the spline should be
evaluated at. Default is 100..}
}
\value{
Returns the \code{n.points} by 2 matrix of evaluations of the B-spline.
}
\description{
Add piecewise B-splines of supplied order to a plot. The function
uses the \code{lines} function to plot to the current device.
}
\examples{
n <- 20
x <- list(x = cumsum(rnorm(2*n)), y = cumsum(rnorm(2*n)))
plot(x, pch = 16, col = "grey")
i <- seq_along(x[[1]] - 1)
with(x, arrows(x[i], y[i], x[i+1], y[i+1], col = "grey"))
bSpline(x, order = 4, col = "red")
bSpline(x, order = 5, col = "blue")
print(bSpline(x, order = 6, col = "green"))
}
\author{
Anders Ellern Bilgrau
}
\seealso{
\code{\link{lines}}
}

