\name{ceiling.lg}

\alias{ceiling.lg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
ceiling largest
}
\description{
Ceiling to largest digit
}
\usage{
ceiling.lg(x)
}

\arguments{
  \item{x }{Numeric vector}

}
\details{
Gives the  ceiling to largest digit (i.e., 54 -> 60).
}

\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
ceiling.lg(250)
ceiling.lg(25000000)
}

