% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char2fac.R
\name{char2fac}
\alias{char2fac}
\title{Convert Non-numeric Columns to Factors}
\usage{
char2fac(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Converts all non-numeric columns in \code{x} to factors.
}
\examples{
d <- data.frame(x = 1:3, 
                y = letters[1:3], 
                z = c(TRUE, TRUE, FALSE), 
                stringsAsFactors = FALSE)
columnTypes(d)
columnTypes(char2fac(d))
}

