% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{check_genes}
\alias{check_genes}
\title{Correct old or Excel damaged HGNC symbols}
\usage{
check_genes(original.symbol, chromosome)
}
\arguments{
\item{original.symbol}{gene symbols to check}

\item{chromosome}{chromosomes for each gene symbol}
}
\value{
vector of predicted gene names based on chromosome info
}
\description{
Correct old or Excel damaged HGNC symbols
}
