% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClassBalance.R
\name{checkClassBalance}
\alias{checkClassBalance}
\title{Check Class Balance}
\usage{
checkClassBalance(y, na.rm = FALSE, ...)
}
\arguments{
\item{y}{A vector of class labels.}

\item{na.rm}{Logical indicating whether or not to remove \code{NA} values
before computation proceeds.}

\item{...}{Additional optional arguments to be passed onto \code{table}.}
}
\value{
Returns the ratio of the frequencies for the largest class and the 
  smallest class.
}
\description{
Calculate the ratio of the most common class to the least common class
}

