% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{checkfun}
\alias{checkfun}
\title{checkfun}
\usage{
checkfun(a, tau)
}
\arguments{
\item{a}{vector to compute quantiles for}

\item{tau}{between 0 and 1, ex. .5 implies get the median}
}
\value{
numeric value
}
\description{
The check function used for optimizing to get quantiles
}
\examples{
x <- rnorm(100)
x[which.min(checkfun(x, 0.5))] ##should be around 0

}
