% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clip.R
\name{clip}
\alias{clip}
\alias{clip.array}
\alias{clip.default}
\alias{clip.matrix}
\title{Clip Values}
\usage{
clip(x, .min, .max, ...)

\method{clip}{default}(x, .min, .max, ...)

\method{clip}{matrix}(x, .min, .max, ...)

\method{clip}{array}(x, .min, .max, ...)
}
\arguments{
\item{x}{A vector, matrix, or multi-way array.}

\item{.min}{.minimum value.}

\item{.max}{.maximum value.}

\item{...}{Additional optional arguments.}
}
\description{
Clip (i.e., limit) the values in a vector, matrix, or array.
}

