% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{col_types}
\alias{col_types}
\title{Columns Types}
\usage{
col_types(x)
}
\arguments{
\item{x}{A data frame or an R object that can be coerced into one.}
}
\description{
Extract the data type of each column of a data frame.
}
