% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_vec.R
\name{col_vec}
\alias{col_vec}
\alias{col_vec.default}
\alias{col_vec.matrix}
\alias{col_vec.data.frame}
\title{Column Vectors}
\usage{
col_vec(..., use.names = TRUE)

\method{col_vec}{default}(..., use.names = TRUE)

\method{col_vec}{matrix}(..., use.names = TRUE)

\method{col_vec}{data.frame}(..., use.names = TRUE)
}
\arguments{
\item{...}{The elements of the columns vector.}

\item{use.names}{Logical indicating whether or not names should be preserved.}
}
\value{
A row vector of dimension \code{length(x) x 1}.
}
\description{
Constructs a column vector.
}
\examples{
col_vec(1:5)
col_vec(row_vec(1:5))

X <- matrix(rnorm(20), ncol = 2)
b <- col_vec(1.5, 3)
(yhat <- X \%*\% b)
}
