% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{color_wheel}
\alias{color_wheel}
\title{Color Wheel}
\usage{
color_wheel(str, ...)
}
\arguments{
\item{str}{A Character string representing the color of interest.}

\item{...}{Additional optional arguments to be passed on to \code{pie}.}
}
\description{
Plot a color wheel for a given color.
}
\examples{
color_wheel("purple")
color_wheel("springgreen")
color_wheel("blue", cex = 0.75)
}

