% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorWheel.R
\name{colorWheel}
\alias{colorWheel}
\title{Color Wheel}
\usage{
colorWheel(str, ...)
}
\arguments{
\item{str}{A Character string representing the color of interest.}

\item{...}{Additional optional arguments to be passed on to \code{pie}.}
}
\description{
Plot a color wheel for a given color.
}
\examples{
colorWheel("purple")
colorWheel("springgreen")
colorWheel("blue", cex = 0.75)
}

