% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num.R, R/summarize_data.R
\name{column_types}
\alias{column_types}
\title{Column Types}
\usage{
column_types(x, ...)

column_types(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{x}{A data frame or an R object that can be coerced into one.}
}
\description{
\code{column_types} returns a vector indicating the type (i.e., class) of
each column of \code{x}.

Extract the data type of each column of a data frame.
}

