% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeData.R
\name{columnTypes}
\alias{columnTypes}
\title{Columns Types}
\usage{
columnTypes(x, ...)
}
\arguments{
\item{x}{A data frame or an R object that can be coerced into one.}
}
\description{
Extract the data type of each column of a data frame.
}

