% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combineAndDraw}
\alias{combineAndDraw}
\title{Draw two oriented graphs with same layout}
\usage{
combineAndDraw(g1, g2, col1 = "green", col2 = "red", name = "", ...)
}
\arguments{
\item{g1}{A graphNEL object}

\item{g2}{A graphNEL object}

\item{col1}{Edge color of edges unique to g1}

\item{col2}{Edge color of edges unique to g2}

\item{name}{The name of the graphs.}

\item{\dots}{Arguments passed to \code{agopen}.}
}
\value{
A \code{list} of length three with the laid out graphs. The first
  and second entry is the laid out g1 and g2 with coloured unique edges.
  The third element is the merged graph (rarely of particular interest).
}
\description{
This function aids in drawing two directed graphNEL objects to have the same
node layout but still preserving the edges. This is useful when comparing
two graphs.
}
\details{
The function works by combining \code{g1} and \code{g2} and making
  unique edges of \code{g2} invisible in the merged graph and vice versa
  for the unique edges of \code{g1}.
}
\examples{
library("gRbase")
g1 <- dagList(list(~A|B, ~A|C, ~A|D, ~E, ~F|A, ~G))
g2 <- dagList(list(~A|B, ~C|A, ~D, ~A|E, ~F|A))
cc <- combineAndDraw(g1, g2)

col1 <- rep("green", length(edgeNames(g1)))
names(col1) <- edgeNames(g1)
col2 <- rep("red", length(edgeNames(g2)))
names(col2) <- edgeNames(g2)

par(oma = c(0,0,0,0)+.2)
layout(rbind(1:2,3,4:5))
plot(g1, edgeAttrs = list(color = col1), main = "Graph 1"); box()
plot(g2, edgeAttrs = list(color = col2), main = "Graph 2"); box()
plot(cc[[3]], main = "Merged graph"); box()
plot(cc[[1]], main = "Graph 1 (laid out as merged graph)"); box()
plot(cc[[2]], main = "Graph 2 (laid out as merged graph)"); box()
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link{agopen}}, \code{\link{plot.graphNEL}}
}

