% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{combineDfs}
\alias{combineDfs}
\title{combineDfs}
\usage{
combineDfs(y.seq, dflist, pstrat = NULL, ...)
}
\arguments{
\item{y.seq}{sequence of possible y values}

\item{...}{additional arguments that can be past to BMisc::make_dist}
}
\description{
Legacy version of `combine_ecdfs`, please use that
 function instead.  This function will eventually be deleted.
}
\keyword{internal}
