% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeWeigtedDiffNetwork.R
\name{computeWeigtedDiffNetwork}
\alias{computeWeigtedDiffNetwork}
\title{Find differential networks between two classes}
\usage{
computeWeigtedDiffNetwork(x, class, gene.universe, name = "datasetname",
  pcor.type = c("marginal", "partial", "fullpartial"), use.TOM = TRUE,
  rho = 0)
}
\arguments{
\item{x}{is the expression matrix where columns correspond to
samples/subjects and rows to genes.}

\item{class}{is the factor encoding ABC and GCB samples}

\item{gene.universe}{character vector giving the names of the features to be
used. I.e. a subset of \code{colnames(x)}.}

\item{name}{A character giving the name of the dataset.}

\item{pcor.type}{character of length one. The type of correlation to be
used. Equal to one of "marginal", "partial", or "fullpartial".}

\item{use.TOM}{logical. Should the topological overlap matrix be computed?
of the P-values of the Fisher transformed tests?}

\item{rho}{Non-negative numeric regularization parameter for lasso. rho = 0
means no regularization. Only used if pcor.type == "fullpartial"}
}
\value{
A list of length 3:
  \item{Z.scores}{A matrix of Z-scores for differential (partial)
    correlation.}
  \item{adjacency}{A matrix of 1 - P-values for the hypothesis of no
    differential (partial) correlation}
  \item{tom.dissimilarity}{A matrix of topological overlap dissimilarities.}
}
\description{
This function attempts to find differential networks between two classes.
First, it computes the partial correlation (either unconditioned, the
smallest first order conditioned, or full conditioned) for each group.
Second it tests each partial correlation for differential correlation using
Fishers Z transform and the normal distribution. If wanted, the topological
overlap matrix is computed from the resulting matrix of P-values.
}
\details{
If \code{pcor.type} is \code{"partial"} then all 1 order partial
  correlations are computed (i.e. the partial correlation given a third
  feature) and the minimal partial correlation is used.
}
\note{
The topological overlap matrix is computed on the P-values.
}

