% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprint.R
\name{cprint}
\alias{cprint}
\alias{cprint.character}
\alias{cprint.data.frame}
\alias{cprint.numeric}
\title{Copy Print}
\usage{
cprint(x, ...)

\method{cprint}{numeric}(x, ...)

\method{cprint}{character}(x, ...)

\method{cprint}{data.frame}(x, ...)
}
\arguments{
\item{x}{An appropriate R object.}

\item{...}{Additional optional arguments. Currently ignored.}
}
\description{
The cprint function provides a capability to print common R data structures 
in a form which can be used as input to the interpreter.
}
\examples{
cprint(iris)  # output can easily be copied into an RStudio script file
}

