% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{cs2panel}
\alias{cs2panel}
\title{Cross Section to Panel}
\usage{
cs2panel(cs1, cs2, yname)
}
\arguments{
\item{cs1}{data frame, the first cross section}

\item{cs2}{data frame, the second cross section}

\item{yname}{the name of the variable to calculate difference for (should be the same in each dataset)}
}
\value{
the change in outcomes over time
}
\description{
Turn repeated cross sections data into panel data by
 imposing rank invariance; does not require
 that the inputs have the same length
}
\examples{
cs1 <- data.frame(y = rnorm(100))
cs2 <- data.frame(y = rnorm(100, mean = 1))
dy <- cs2panel(cs1, cs2, "y")
mean(dy) ## approx 1

}
