% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical indicating whether or not \code{NA} values should be 
stripped before the computation proceeds.}
}
\description{
Calculate the sample coefficient of variation for a vector of data.
}
\examples{
cv(rnorm(100))
}
