% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cv_tree.R
\name{cv_tree}
\alias{cv_tree}
\title{K-Fold Cross-Validation Tree}
\usage{
cv_tree(x, y, cv.folds = 10, tune.length = 10)
}
\arguments{
\item{x}{A matrix or predictor variables.}

\item{y}{A vector of response values.}

\item{cv.folds}{In teger representing the number of folds to use for cross-
validation.}
}
\description{
Build a decision tree based on k-fold cross-validation.
}
\examples{
data(mushroom, package = "misc")
tree <- cv_tree(x = subset(mushroom, select = -Edibility),
                y = mushroom$Edibility)
plot(tree)
}

