% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.cor.test.R
\name{differential.cor.test}
\alias{differential.cor.test}
\title{Test for differential correlation}
\usage{
differential.cor.test(cor1, cor2, N1, N2)
}
\arguments{
\item{cor1}{A numeric matrix of correlation coefficients in the first group.}

\item{cor2}{As c\code{cor1} for the second group.}

\item{N1}{integer of lenth 1. The number of samples in group 1.}

\item{N2}{integer of lenth 1. The number of samples in group 2.}
}
\value{
A list of matrices or vector containing:
  \item{\code{Z.scores}}{A numeric matrix of Z-scores for the hypothesis.}
  \item{\code{P.values}}{A numeric matrix of the P-values.}
}
\description{
This functions tests the hypothesis of no difference in correlations.
}
\details{
The function uses Fisher's Z transform (atanh) of correlations
}
\examples{
N1 <- 10
print(cor1 <- cor(t(replicate(N1, rnorm(5)))))
N2 <- 8
print(cor2 <- cor(t(replicate(N2, rnorm(5)))))

differential.cor.test(cor1, cor2, N1, N2)
}
\references{
\url{http://core.ecu.edu/psyc/wuenschk/docs30/CompareCorrCoeff.pdf}
}
