% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{drop_collinear}
\alias{drop_collinear}
\title{drop_collinear}
\usage{
drop_collinear(matrix)
}
\arguments{
\item{matrix}{a matrix for which the function will remove collinear columns}
}
\value{
a matrix with collinear columns removed
}
\description{
A function to check for multicollinearity and drop collinear terms
 from a matrix
}
\examples{
\donttest{
  if (requireNamespace("caret", quietly = TRUE)) {
    X <- cbind(1:5, 2 * (1:5), rnorm(5))
    colnames(X) <- c("x1", "x2", "x3")
    drop_collinear(X) ## x2 dropped as collinear with x1
  }
}
}
