% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{drop_cov_from_formula}
\alias{drop_cov_from_formula}
\title{Drop a Covariate from a Formula}
\usage{
drop_cov_from_formula(covs, formula)
}
\arguments{
\item{covs}{should be a list of variable names}

\item{formla}{which formula to drop covariates from}
}
\value{
formula
}
\description{
\code{drop_cov_from_formula} adds drops some covariates from a
formula; covs should be a list of variable names
}
\examples{
ff <- y ~ x + w + z
drop_cov_from_formula(list("w", "z"), ff)

drop_cov_from_formula("z", ff)

}
