% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{dropCovFromFormla}
\alias{dropCovFromFormla}
\title{dropCovFromFormla}
\usage{
dropCovFromFormla(covs, formla)
}
\arguments{
\item{covs}{should be a list of variable names}
}
\description{
Legacy version of `drop_cov_from_formula`, please use
 that function instead.  This function will eventually be deleted.
}
\keyword{internal}
