% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{extract_aa}
\alias{extract_aa}
\title{Extract Amino Acid Variant}
\usage{
extract_aa(aachange, type)
}
\arguments{
\item{aachange}{string containing protein sequence variant}

\item{type}{string indicating whether to extract the reference or alternate
variant. Use 'ref' to indicate reference and 'alt' to indicate the alternate
variant.}
}
\value{
amino acid variant specified by type
}
\description{
Extract Amino Acid Variant
}
