% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{extract_position}
\alias{extract_position}
\title{Extract Amino Acid Variant Position}
\usage{
extract_position(aachange)
}
\arguments{
\item{aachange}{string containing protein sequence variant}
}
\value{
amino acid variant's position
}
\description{
Extract Amino Acid Variant Position
}
