% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num.R
\name{factorize}
\alias{factorize}
\title{Convert Non-numeric Columns}
\usage{
factorize(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
\code{factorize} converts all non-numeric columns in \code{x} to factors.
}
\examples{
d <- data.frame(x = 1:3,
                y = letters[1:3],
                z = c(T, T, F),
                stringsAsFactors = FALSE)
column_types(d)
column_types(factorize(d))
str(d)
str(factorize(d))
}

