\name{fct}

\alias{fct}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print bmisc functions
}
\description{
Prints all functions of bmisc package
}
\usage{
fct()
}
%- maybe also 'usage' for other objects documented here.
\details{

}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}

