% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_california_housing.R
\name{fetch_california_housing}
\alias{fetch_california_housing}
\title{Load California housing data}
\source{
https://scikit-learn.org/stable/modules/generated/sklearn.datasets.fetch_california_housing.html
}
\usage{
fetch_california_housing(raw = FALSE, tibble = FALSE)
}
\arguments{
\item{raw}{Logical indicating whether or not to return the raw data or a 
cleaned up version. Default is \code{FALSE}.}

\item{tibble}{Logical indicating whether or not to return the results as a 
\code{\link[tibble]{tibble}}, as opposed to a data frame. Default is 
\code{FALSE}.}
}
\description{
Load California housing data from source.
}
\examples{
\dontrun{
cal <- fetch_california_housing(tibble = TRUE)
}
}
