% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strings.R
\name{first_alpha}
\alias{first_alpha}
\title{String Methods}
\usage{
first_alpha(string)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}
}
\value{
Returns the first alphabetic character (A-Z or a-z) or \code{NA} if
  no alphabetic characters are found.
}
\description{
Extract the first alphabetic character (if available) from the characters of
a string.
}

