% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strings.R
\name{first_digit}
\alias{first_digit}
\title{String Methods}
\usage{
first_digit(string)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}
}
\value{
Returns the first digit (0-9) or \code{NA} if no digits are found.
}
\description{
Extract the first digit (if available) from the characters of a string.
}

