% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip names and values}
\usage{
flip(x)
}
\arguments{
\item{x}{A named vector.}
}
\value{
A character vector.
}
\description{
Flips or switches names and values of vectors.
}
\examples{
x <- structure(1:5, names = LETTERS[1:5])
x
flip(x)
flip(flip(x))
}
\author{
Anders Ellern Bilgrau
}

