% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatForStackOverflow.R
\name{formatForStackOverflow}
\alias{formatForStackOverflow}
\title{Run and format R code to StackOverflow submission ready}
\usage{
formatForStackOverflow(user_code, comment_eval = TRUE, file = "")
}
\arguments{
\item{user_code}{A character of length 1 of user code to run and format.}

\item{comment_eval}{logical. If \code{TRUE} output is prefixed with a comment.}

\item{file}{A path to a file to write the output to. The default prints to console.}
}
\value{
Prints to console or file a ready-to-copy to Stack Overflow formatted code and output.
}
\description{
Runs the R code, captures both code and output and prints such 
that code and output are readily copied into an Stack Overflow post or answer.
I.e. it prefixes four spaces and (optinally) comments output.
}
\examples{
formatForStackOverflow("
  myfunc <- function(x) {
    return(c(mean=mean(x), med=median(x)))
  }
  
  # A comment
  n <- 100
  myfunc(rnorm(n))
  
  sapply(mtcars, myfunc) # inline comment
  
  print(mtcars)
")
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
