% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.randomForest}
\alias{fortify.randomForest}
\title{Random Forest Summary Data}
\usage{
\method{fortify}{randomForest}(object, values = c("importance", "oob-error",
  "oob-rsquared", "margin", "parallel", "partial", "proximity-heat",
  "proximity-mds", "proximity-outlier"), fac, ...)
}
\arguments{
\item{object}{An object of class \code{"randomForest"}.}

\item{values}{The type of data to collect.}

\item{fac}{A factor that was used as the response to train the random forest.
Only used if \code{values = "proximity-mds"} or
\code{values = "proximity-outlier"}.}

\item{...}{Additional arguments.}
}
\value{
A data frame.
}
\description{
Extract information from a \code{"randomForest"} object into a data frame.
}

