% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.attribute}
\alias{get.attribute}
\title{Convert from bioconductor database array name to biomaRt's array names}
\usage{
get.attribute(es)
}
\arguments{
\item{es}{is an ExpressionSet object}
}
\value{
A character of length \code{1} with the biomaRt array name
}
\description{
Convert from bioconductor database array name to biomaRt's array names
}
\note{
Note,
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

