% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.lower.tri-get.upper.tri.R
\name{get.lower.tri}
\alias{get.lower.tri}
\alias{get.lower.tri<-}
\alias{get.upper.tri}
\alias{get.upper.tri<-}
\title{Vectorize the upper and lower triangular matrix}
\usage{
get.lower.tri(x, diag = FALSE)

get.lower.tri(x, diag = FALSE) <- value

get.upper.tri(x, diag = FALSE)

get.upper.tri(x, diag = FALSE) <- value
}
\arguments{
\item{x}{A \code{matrix}.}

\item{diag}{logical. Should the diagonal be included?}

\item{value}{A vector of the same data type a \code{x} of the same length
as \code{get.lower.tri}.}
}
\value{
\code{get.lower.tri} returns a numeric vector of the entries in the
  below the diagonal of \code{x}. The length is dependent on argument
  \code{diag}.

\code{get.upper.tri} returns a numeric vector of the entries in the
  below the diagonal of \code{x}. The length is dependent on argument
  \code{diag}.
}
\description{
Get and set the upper and lower triangular matrix in vector form.
}
\details{
The returned length is dependent on \code{diag}. If
  \code{diag == TRUE}, then the length is \code{ncol(x)*(ncol(x)-1)/2}. If
  \code{FALSE}, then the length is \code{ncol(x)*(ncol(x)+1)/2}.
}
\examples{
print(x <- matrix(1:25, ncol = 5))
get.lower.tri(x)
get.lower.tri(x, diag = TRUE)

print(y <- matrix(1:9, ncol = 3))
get.upper.tri(y)
get.upper.tri(y, diag = TRUE)

get.upper.tri(y) <- get.lower.tri(y)  # Fill the upper tri with the lower
print(y)
get.upper.tri(y)

print(x)
get.lower.tri(x, diag = TRUE) <- get.upper.tri(x, diag = TRUE)
print(x)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
\seealso{
\code{\link{lower.tri}}, \code{\link{upper.tri}}
}

