% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.lower.tri}
\alias{get.lower.tri}
\alias{get.upper.tri}
\title{Vectorize the upper and lower triangluar matrix}
\usage{
get.lower.tri(x, diag = FALSE)

get.upper.tri(x, diag = FALSE)
}
\arguments{
\item{x}{A numeric matrix.}

\item{diag}{logical. Should the diagonal be included?}
}
\value{
\code{get.lower.tri} returns a numeric vector of the entries in the
  below the diagonal of \code{x}. The length is dependent on argument
  \code{diag}.

\code{get.upper.tri} returns a numeric vector of the entries in the
  below the diagonal of \code{x}. The length is dependent on argument
  \code{diag}.
}
\description{
Vectorize the upper and lower triangluar matrix
}
\details{
The returned length is dependent on \code{diag}. If
  \code{diag == TRUE}, then the length is \code{ncol(x)*(ncol(x)-1)/2}. If
  \code{FALSE}, then the length is \code{ncol(x)*(ncol(x)+1)/2}.
}
\examples{
print(x <- matrix(1:25, ncol = 5))
get.lower.tri(x)
get.lower.tri(x, diag = TRUE)
print(x <- matrix(1:25, ncol = 3))
get.upper.tri(x)
get.upper.tri(x, diag = TRUE)
}
\author{
Anders Ellern Bilgrau (abilgrau (at) math.aau.dk)
}
\seealso{
\link{\code{lower.tri}}, \link{\code{upper.tri}}
}

