% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{get_group}
\alias{get_group}
\title{get_group}
\usage{
get_group(df, idname, tname, treatname)
}
\arguments{
\item{df}{the data.frame used in the function}

\item{idname}{name of column that holds the unit id}

\item{tname}{name of column that holds the time period}

\item{treatname}{name of column with the treatment indicator}
}
\description{
A function to calculate a unit's group in a panel data setting
 with a binary treatment and staggered treatment adoption and where
 there is a column in the data indicating whether or not a unit is treated
}
\examples{
n <- 50
id <- rep(seq_len(n), each = 4)
t <- rep(1:4, n)
g <- rep(sample(c(0, 2, 3), n, replace = TRUE), each = 4)
treat <- as.integer(t >= g & g > 0)
dta <- data.frame(id = id, t = t, treat = treat)
dta$group <- get_group(dta, idname = "id", tname = "t", treatname = "treat")
head(unique(dta[, c("id", "group")]))
}
