% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{get_group_inner}
\alias{get_group_inner}
\title{get_group_inner}
\usage{
get_group_inner(this_df, tname, treatname)
}
\arguments{
\item{this_df}{a data.frame, for this function it should be specific to
a particular unit}

\item{tname}{name of column that holds the time period}

\item{treatname}{name of column with the treatment indicator}
}
\description{
Calculates the group for a particular unit
}
\keyword{internal}
