% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{get_list_element}
\alias{get_list_element}
\title{Return Particular Element from Each Element in a List}
\usage{
get_list_element(listolists, whichone = 1)
}
\arguments{
\item{listolists}{a list}

\item{whichone}{which item to get out of each list (can be numeric or name)}
}
\value{
list of all the elements 'whichone' from each list
}
\description{
a function to take a list and get a particular part
 out of each element in the list
}
\examples{
len <- 100 # number elements in list
lis <- lapply(1:len, function(l) list(x = (-l), y = l^2)) # create list
get_list_element(lis, "x")[1] # should be equal to -1
get_list_element(lis, 1)[1] # should be equal to -1

}
