% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{get_Yibar}
\alias{get_Yibar}
\title{get_Yibar}
\usage{
get_Yibar(df, idname, yname)
}
\arguments{
\item{df}{the data.frame used in the function}

\item{idname}{name of column that holds the unit id}

\item{yname}{name of column containing the outcome (or other variable)
for which to calculate its outcome in the immediate pre-treatment period}
}
\description{
A function to calculate the average outcome across all time
periods separately for each unit in a panel data setting (this function can also
 be used to recover covariates, etc.).
}
\examples{
n <- 50
id <- rep(seq_len(n), each = 4)
t <- rep(1:4, n)
y <- rnorm(n * 4)
dta <- data.frame(id = id, t = t, y = y)
dta$Yibar <- get_Yibar(dta, idname = "id", yname = "y")
}
