% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{get_Yit_inner}
\alias{get_Yit_inner}
\title{get_Yit_inner}
\usage{
get_Yit_inner(this_df, tp, yname, tname)
}
\arguments{
\item{this_df}{a data.frame, for this function it should be specific to
a particular unit}

\item{tp}{The time period for which to get the outcome}

\item{yname}{name of column containing the outcome (or other variable)
for which to calculate its outcome in the immediate pre-treatment period}

\item{tname}{name of column that holds the time period}
}
\description{
Calculates a units outcome in some particular period `tp`.
 This function operates on a data.frame that is already local to a particular
 unit.
}
\keyword{internal}
