% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{getListElement}
\alias{getListElement}
\title{getListElement}
\usage{
getListElement(listolists, whichone = 1)
}
\arguments{
\item{listolists}{a list}

\item{whichone}{which item to get out of each list (can be numeric or name)}
}
\description{
Legacy version of `get_list_element`, please use that function
 instead.  This function will eventually be deleted.
}
\keyword{internal}
