% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggErrorPlot}
\alias{ggErrorPlot}
\title{Plot of Random Forest OOB Error}
\usage{
ggErrorPlot(object, include.legend = TRUE, xlab = "Number of trees",
  ylab = "OOB error", main = "Training Error Plot\\n", ...)
}
\arguments{
\item{object}{An object of class \code{"randomForest"}.}

\item{include.legend}{Logical indicating whether to include a legend. Default
is \code{TRUE}.}

\item{xlab}{A label for the x axis.}

\item{ylab}{A label for the y axis.}

\item{main}{A main title for the plot.}
}
\description{
ggplot2-based lineplot of oob-based error for a random forest object.
}

