% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggMarginPlot}
\alias{ggMarginPlot}
\title{Margin Plot}
\usage{
ggMarginPlot(object, include.legend = TRUE, xlab = "Index",
  ylab = "Margin", main = "Margin Plot\\n", ...)
}
\arguments{
\item{object}{A \code{randomForest} object.}

\item{include.legend}{Logical indicating whether a legend should be drawn
on the plot.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{main}{Plot title.}

\item{...}{Additional optional arguments.}
}
\description{
Margin Plot
}

