% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggParallelCoordinatePlot}
\alias{ggParallelCoordinatePlot}
\title{Random Forest Parallel Coordinates Plot}
\usage{
ggParallelCoordinatePlot(object, include.legend = FALSE, xlab = "Variable",
  ylab = "Importance score", main = "Parallel Coordinate Plot\\n", ...)
}
\arguments{
\item{object}{#'}

\item{include.legend}{Logical indicating whether to include a legend. Default
is \code{TRUE}.}

\item{xlab}{A label for the x axis.}

\item{ylab}{A label for the y axis.}

\item{main}{A main title for the plot.}

\item{...}{Additional optional arguments.}

\item{object}{An object of class \code{"randomForest"}.}
}
\description{
Parallel coordinates plot of random forest feature importance scores for each
class.
}
\examples{
# Forensic glass data
data(Glass, package = "mlbench")
rf_glass <- randomForest(Type ~ ., data = Glass, importance = TRUE)
ggParallelCoordinatePlot(rf_glass, include.legend = TRUE)
}

