% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggPartialPlot}
\alias{ggPartialPlot}
\title{Random Forest Partial Dependence Plot}
\usage{
ggPartialPlot(object, smooth = TRUE, ...,
  main = "Partial Dependence Plot\\n", xlab = list(...)$x.var,
  ylab = "Partial dependence")
}
\arguments{
\item{object}{An object of class \code{"randomForest"}.}

\item{smooth}{Logical indicating whether to add a scatterplot smoother to the
plot. Default is \code{TRUE}.}

\item{...}{Additional arguments required by \code{partialPlot}. AT a minimum,
\code{x.var} and \code{pred.data} should be given, althouh here,
\code{x.data} should be a character vector giving the name of the predictor
variable of interest.}

\item{main}{A main title for the plot.}

\item{xlab}{A label for the x axis.}

\item{ylab}{A label for the y axis.}
}
\description{
ggplot2 version of \code{randomForest::partialPlot}.
}
\examples{
# Boston housing data (regression)
data(Boston, package = "MASS")
rf_boston <- randomForest(medv ~ ., data = Boston)
ggPartialPlot(rf_boston, pred.data = Boston, x.var = "rm")

# Iris data (classification)
data(iris, package = "datasets")
rf_iris <- randomForest(Species ~ ., data = iris)
ggPartialPlot(rf_iris, pred.data = iris, x.var = "Petal.Width")
}

