% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggProximityPlot}
\alias{ggProximityPlot}
\title{Random Forest Proximity Matrix Plot}
\usage{
ggProximityPlot(object, plot.type = c("heat", "MDS", "outlier"), fac,
  size = 3, alpha = 1, include.legend = TRUE, ..., xlab = if (plot.type
  == "MDS") "Dimension 1" else "", ylab = if (plot.type == "MDS")
  "Dimension 2" else "", main = if (plot.type == "MDS")
  "MDS Plot of Proximity Matrix\\n" else "Heat Map of Proximity Matrix\\n")
}
\arguments{
\item{object}{An object of class \code{"randomForest"}.}

\item{plot.type}{Type of plot to produce. Current options are
\code{"heat"} for a heat map of the proximity matrix or \code{"MDS"} for a
multi-dimensional scaling plot of the proximity matrix.}

\item{include.legend}{Logical indicating whether to include a legend. Default}

\item{...}{Additional optional arguments. This option is currently ignored.}

\item{xlab}{A label for the x axis.}

\item{ylab}{A label for the y axis.}

\item{main}{A main title for the plot.}
}
\description{
ggplot2 version of \code{randomForest::MDSplot} and more.
}
\examples{
# Motor trend car road test data (regression)
data(mtcars, package = "datasets")
rf_mtcars <- randomForest(mpg ~ ., data = mtcars, proximity = TRUE)
ggProximityPlot(rf_mtcars, plot.type = "heat")

# Iris data (classification)
data(iris, package = "datasets")
rf_iris <- randomForest(Species ~ ., data = iris, proximity = TRUE)
ggProximityPlot(rf_iris, plot.type = "MDS", fac = iris$Species, size = 5,
                alpha = 0.5)
}

