% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggVarImpPlot}
\alias{ggVarImpPlot}
\title{Random Forest Variable Importance Plot}
\usage{
ggVarImpPlot(object, type = NULL, class = NULL, scale = TRUE, ...,
  main = "Feature Importance Plot\\n", xlab = "",
  ylab = "Importance score")
}
\arguments{
\item{object}{An object of class \code{"randomForest"}.}

\item{type}{Either 1 or 2, specifying the type of importance measure
(1=mean decrease in accuracy, 2=mean decrease in node impurity).}

\item{class}{For classification problem, which class-specific measure to
return.}

\item{scale}{For permutation based measures, should the measures be divided
  their "standard errors"?
... Additional optional arguments. This option is currently ignored.}

\item{...}{Additional optional arguments.}

\item{main}{A main title for the plot.}

\item{xlab}{A label for the x axis.}

\item{ylab}{A label for the y axis.}
}
\description{
ggplot2 version of \code{randomForest::varImpPlot}.
}

