% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/math.R
\name{hypot}
\alias{hypot}
\title{Miscelaneous math functions}
\usage{
hypot(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\value{
The length of the hypotenuse.
}
\description{
Calculate the length of the hypotenuse of a right-angle triangle.
}

