% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{id2rownum}
\alias{id2rownum}
\title{Take particular id and convert to row number}
\usage{
id2rownum(id, data, idname)
}
\arguments{
\item{data}{data frame}

\item{idname}{unique id}

\item{ids}{vector of ids}
}
\description{
id2rownum takes an id and converts it t the right
 row number in the dataset; ids should be unique in the dataset
 that is, don't pass the function panel data with multiple same ids
}
\keyword{internal}
