% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importFrom.R
\name{importFrom}
\alias{importFrom}
\title{Import Specific Objects from Packages}
\usage{
importFrom(pkg, objects, envir = .GlobalEnv)
}
\arguments{
\item{pkg}{An R package.}

\item{objects}{Character vector specifying the particular objects to import
from \code{pkg}.}

\item{envir}{The environment for which the objects will be loaded into. The
default is \code{.GlobalEnv} meaning load the objects into the global
environment.}

\item{names}{Optional character vector speccifying the name assigned to the
  imported objects. Currently not implemented.}
}
\description{
Import specific objects (e.g., functions) from a packages namespace.
}
\details{
Normally, one could implement this via \code{object <- pkg::object}, but that
would require a new assignment for each object imported. This method allows
one to import many objects at once from a particular package.
}

