% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect a Data Frame}
\usage{
inspect(x, n = 8)
}
\arguments{
\item{x}{A data frame.}

\item{n}{Integer specifying the number of rows to inspect at random.}
}
\description{
Inspect a data frame.
}

