% Generated by roxygen2 (4.0.1): do not edit by hand
\name{install.brainarray}
\alias{install.brainarray}
\title{Installs brainarray CDF and probe information packages}
\usage{
install.brainarray(array, version, type = "ensg", force.reinstall = FALSE,
  force.download = FALSE, use.temp.dir = TRUE, path)
}
\description{
Installs brainarray CDF and probe information packages
}

