% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{invert_ecdf}
\alias{invert_ecdf}
\title{Invert Ecdf}
\usage{
invert_ecdf(df)
}
\arguments{
\item{df}{an ecdf object}
}
\value{
stepfun object that contains the quantiles of the df
}
\description{
take an ecdf object and invert it to get a step-quantile
 function
}
\examples{
y <- rnorm(100)
F <- ecdf(y)
Finv <- invert_ecdf(F)
Finv(0.5) ## approximate median

}
