% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{invertEcdf}
\alias{invertEcdf}
\title{invertEcdf}
\usage{
invertEcdf(df)
}
\arguments{
\item{df}{an ecdf object}
}
\description{
Legacy function for `invert_ecdf`, please use that
 function instead.  This function will eventually be deleted.
}
\keyword{internal}
