\name{is.even}

\alias{is.even}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
is even
}
\description{
Identifies if a value is even or not
}
\usage{
is.even(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{numeric vector}
}
\details{
Will returns TRUE if \code{\link{roundup}}(x) is an even number.
}
\value{
logical
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{is.odd}}
}
\examples{
is.even(5)
is.even(6)
}

