\name{is.odd}

\alias{is.odd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
is odd
}
\description{
Identifies if a value is odd or not
}
\usage{
is.odd(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{numeric vector}
}
\details{
Will returns TRUE if \code{\link{roundup}}(x) is an odd number.
}
\value{
logical
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{is.even}}
}
\examples{
is.odd(5)
is.odd(6)
}

