% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{is_alnum}
\alias{is_alnum}
\title{String methods}
\usage{
is_alnum(string)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}
}
\value{
Returns \code{TRUE} if all characters in the string are alphanumeric 
and there is at least one character, FALSE otherwise.
}
\description{
General test if characters in a string are alphanumeric.
}
\examples{
is_alnum("abc")
is_alnum("abc1")
is_alnum("1")
}
