% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{is_digit}
\alias{is_digit}
\title{String methods}
\usage{
is_digit(string)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}
}
\value{
Returns \code{TRUE} if all characters in the string are digits and 
there is at least one character, FALSE otherwise.
}
\description{
General test if characters in a string are digits.
}
\examples{
is_digit("abc")
is_digit("abc1")
is_digit("1")
is_digit("1.3")
}
