% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R, R/trees.R
\name{is_root}
\alias{is_root}
\title{Is the Tree a Root?}
\usage{
is_root(object)

is_root(object)
}
\arguments{
\item{object}{An object that inherits from class \code{"rpart"}.}

\item{object}{An object that inherits from class \code{"rpart"}.}

\item{...}{Additional optional arguments. None are currently used.}
}
\description{
\code{is_root} tests if an \code{"rpart"} object is just a root (i.e., no
splits).

Test if a fitted  \code{"rpart"} tree is just a root node (i.e., no splits).
}
\keyword{internal}

