% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{kth.alpha}
\alias{kth.alpha}
\title{String Methods}
\usage{
kth.alpha(string, k = 1L)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}

\item{k}{Integer specifying which alphabetic character to extract.}
}
\value{
Returns the k-th alphabetic character (A-Z or a-z) or \code{NA} if 
  at least \code{k} alphabetic characters are not found.
}
\description{
Extract the k-th alphabetic character (if available) from the characters of 
a string.
}

