% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{kth_alpha}
\alias{kth_alpha}
\title{String methods}
\usage{
kth_alpha(string, k = 1L)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}

\item{k}{Integer specifying which alphabetic character to extract.}
}
\value{
Returns the k-th alphabetic character (A-Z or a-z) or \code{NA} if 
at least \code{k} alphabetic characters are not found.
}
\description{
Extract the k-th alphabetic character (if available) from the characters of 
a string.
}
\examples{
kth_digit("abc0def7ghi5")
kth_digit("abc0def7ghi5", k = 2)
}
