% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strings.R
\name{kth_digit}
\alias{kth_digit}
\title{String Methods}
\usage{
kth_digit(string, k = 1L)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}

\item{k}{Integer specifying which digit to extract.}
}
\value{
Returns the k-th digit (0-9) or \code{NA} if at least \code{k} digits
  are not found.
}
\description{
Extract the k-th digit (if available) from the characters of a string.
}

