% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{last.alpha}
\alias{last.alpha}
\title{String Methods}
\usage{
last.alpha(string)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}
}
\value{
Returns the last alphabetic character (A-Z or a-z) or \code{NA} if 
  no alphabetic characters are found.
}
\description{
Extract the last alphabetic character (if available) from the characters of 
a string.
}

