% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{last.digit}
\alias{last.digit}
\title{String Methods}
\usage{
last.digit(string)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}
}
\value{
Returns the last digit (0-9) or \code{NA} if no digits are found.
}
\description{
Extract the last digit (if available) from the characters of a string.
}

