% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{lhs.vars}
\alias{lhs.vars}
\title{lhs.vars}
\usage{
lhs.vars(formla)
}
\description{
Legacy version of `lhs_vars`, please use that function
 instead.  This function will eventually be deleted.
}
\keyword{internal}
