% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Intersect}
\alias{Intersect}
\alias{ListSetOperations}
\alias{Setdiff}
\alias{Union}
\title{Set operations on lists of character vectors.}
\usage{
Intersect(x)

Setdiff(x, y)

Union(x)
}
\arguments{
\item{x}{A \code{list} of \code{character} vectors.}

\item{y}{A \code{list} of \code{character} vectors.}
}
\value{
\code{Intersect} returns the intersection of all vectors in \code{x}.

\code{Setdiff} returns the intersection of \code{x} minus the
  union of \code{y}.

\code{Union} returns the union all character vectors in \code{x}.
}
\description{
Set operations on lists of character vectors.
}
\details{
These functions perform basic set operations on lists of character
  vectors in a recursive manner. The work analogous to \code{union},
  \code{intersect}, and \code{setdiff} but on lists.
}
\examples{
set.seed(1)
x <- lapply(2:6, function(i) sample(LETTERS, 10+i))
y <- lapply(1:3, function(i) sample(LETTERS, 2+i))

Union(x)  # The union of all vectors in x
Intersect(x)  # The intersection of all vectors in x
Setdiff(x, y)  # All elements in the intersection of x, but not in any y
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link{union}}, \code{\link{intersect}}, \code{\link{setdiff}}
}

