% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{load_packages}
\alias{load_packages}
\alias{unload_packages}
\title{Load/unload multiple packages}
\usage{
load_packages(pkgs)

unload_packages(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names.}
}
\description{
Load or unload multiple packages at once, rather than having to call
\code{library} or \code{detach} separately for each.
}
