% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_utils.R
\name{load.cohorts}
\alias{load.cohorts}
\title{Load multiple TCGA studies at once}
\usage{
load.cohorts(cancers, data.dir, genes)
}
\arguments{
\item{cancers}{TCGA studies (use acronym i.e. SKCM)}

\item{data.dir}{Directory where the data files are stored}

\item{genes}{Genes to include in dataframe. NULL loads all genes}
}
\value{
Dataframe where rows are patients and columns include clinical, rnaseq, and
cnv data. The `cohort`` column contains the TCGA study each patient belongs to.
}
\description{
Load multiple TCGA studies at once
}
