% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_utils.R
\name{load_cnv}
\alias{load_cnv}
\title{Load dataframe with Gistic2 thresholded CNV data}
\usage{
load_cnv(cancer, data.dir, genes = NULL, file = NULL)
}
\arguments{
\item{cancer}{TCGA study (use acronym i.e. SKCM)}

\item{data.dir}{Directory where the data files are stored}

\item{genes}{Genes to include in dataframe. NULL loads all genes}

\item{file}{Filename of the file containing CNV data}
}
\value{
Dataframe where each row is one sample and each column is
the Gistic2 score for that gene
}
\description{
Load dataframe with Gistic2 thresholded CNV data
}
