% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_utils.R
\name{load_cohort}
\alias{load_cohort}
\title{Load dataframe clinical info and optional RNASeq and CNV data}
\usage{
load_cohort(cancer, data.dir, genes = NULL, include.cnv = TRUE,
  include.rna = TRUE, file = NULL)
}
\arguments{
\item{cancer}{TCGA study (use acronym i.e. SKCM)}

\item{data.dir}{Directory where the data files are stored}

\item{genes}{Genes to include in dataframe. NULL loads all genes}

\item{include.cnv}{Whether to load copy number data}

\item{include.rna}{Whether to load rnaseq data}

\item{file}{Filename of the file containing clinical data}
}
\value{
Dataframe where each row is a patient sample with columns containing clinical
info as well as optional RNASeq/CNV data. All patients have complete Overall Survival
data. Solid Tissue Normal samples are not included.
}
\description{
Load dataframe clinical info and optional RNASeq and CNV data
}
