% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{makeBalancedPanel}
\alias{makeBalancedPanel}
\title{makeBalancedPanel}
\usage{
makeBalancedPanel(data, idname, tname)
}
\arguments{
\item{data}{data.frame (or data.table) used in function}

\item{idname}{unique id}

\item{tname}{time period name}
}
\description{
Legacy version of `make_balanced_panel`,
 please use that function name going forward, though this will
 still work for now.
}
\keyword{internal}
