% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mat}
\alias{mat}
\alias{mat.character}
\alias{mat.list}
\title{NumPy style matrix function}
\usage{
mat(x, ...)

\method{mat}{character}(x)

\method{mat}{list}(x, rows = TRUE)
}
\description{
Returns a matrix from a character vector or list of vectors.
}
\examples{
mat('1, 2, 3; 4, 5, 6')
mat(paste('exp(', 1:3, ')'))
}

