% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/minimum.abs.pcor.R
\name{minimum.abs.pcor}
\alias{minimum.abs.pcor}
\title{The minimum absolute first order partial correlation}
\usage{
minimum.abs.pcor(S)
}
\arguments{
\item{S}{A numeric covariancematrix}
}
\value{
A numeric matrix.
}
\description{
This functions computes all first order partial correlations and retuns
the partial correlation with the least absolute size.
}
\examples{
X <- replicate(3, rnorm(5))
dimnames(X) <- list(paste0("obs", 1:nrow(X)), paste0("var", 1:ncol(X)))
S <- cor(X)
minimum.abs.pcor(S)
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link{pcor}}
}

