% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{most_freq}
\alias{most_freq}
\title{Most Frequent Value}
\usage{
most_freq(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
\code{most_freq} returns the most frequent value in \code{x}.
}
\examples{
most_freq(sample(c("a", "b", "c"), size = 100, replace = TRUE, 
                 prob = c(0.1, 0.3, 0.6)))
}
