% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiplier_bootstrap_parallel}
\alias{multiplier_bootstrap_parallel}
\title{multiplier_bootstrap_parallel}
\usage{
multiplier_bootstrap_parallel(inf_func, biters)
}
\arguments{
\item{inf_func}{nxk matrix of (e.g., these could be a matrix
containing the influence function for different parameter
estimates)}

\item{biters}{the number of bootstrap iterations}
}
\value{
a Bxk matrix
}
\description{
A function that takes in an influence function (an
nxk matrix) and the number of bootstrap iterations and
returns a Bxk matrix of bootstrap results. This function
uses Rademechar weights.
}
