% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gammap}
\alias{Dgammap}
\alias{MultivariateSpecial}
\alias{digammap}
\alias{gammap}
\alias{ldigammap}
\alias{lgammap}
\title{Multivariate special functions}
\usage{
gammap(x, p = 1)

lgammap(x, p = 1)

digammap(x, p = 1)

Dgammap(x, p = 1)
}
\arguments{
\item{x}{A numeric vector of values. The values of \code{x} should be
strictly greater than \code{(p - 1)/2}.}

\item{p}{An integer giving the dimension. Default is \code{1}.}
}
\value{
Returns a numeric vector with the same length as \code{x}.
  Similar to \code{\link{gamma}} and related functions \code{NaN}s are
  returned with warning when the function is evaluated outside its domain

  \code{gammap} is the multivariate gamma function.

\code{lgammap} is the log multivariate gamma function.

\code{digammap} is the multivariate digamma function.

\code{Dgammap} is the derivative of the multivariate gamma function.
}
\description{
Special matematical functions related to the multivariate gamma function
which appears in the Wishart and Inverse Wishart distributions. The interface
and usage should mirror the \link{Special} mathematical functions from the
base package.
}
\details{
Warnings are thown when the functions are evaluated outside its domain.
  Run \code{Bmisc:::plotMultivariateSpecial()} for more info.
}
\examples{
gammap(3.2, p = 1)
gammap(c(0.2, 0.5, 0.9, 1.2, 1.5), p = 3)
lgammap(3.1, p = 2)
digammap(2, p = 3)
Dgammap(3.2, p = 1)

# A plot of all special functions
Bmisc:::plotMultivariateSpecial()
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
The \link{Special} mathematical functions.
}

