% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mushroom.R
\docType{data}
\name{mushroom}
\alias{mushroom}
\title{Mushrooms}
\format{A data frame with 8124 rows and 23 variables.}
\source{
Knopf, Alfred A. \emph{The Audubon Society Field Guide to North American
Mushrooms}, G. H. Lincoff (Pres.), 1981.
}
\description{
Mushrooms described in terms of physical characteristics.
}
\keyword{datasets}

