% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{mv_mult}
\alias{mv_mult}
\title{Matrix-Vector Multiplication}
\usage{
mv_mult(A, v)
}
\arguments{
\item{A}{an nxk matrix.}

\item{v}{a vector (can be stored as numeric or as a kx1 matrix)}
}
\value{
A numeric vector resulting from the multiplication of the matrix by the vector.
}
\description{
This function multiplies a matrix by a vector and returns a numeric vector.
}
\examples{
A <- matrix(1:9, nrow = 3, ncol = 3)
v <- c(2, 4, 6)
mv_mult(A, v)
}
