% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hybrid.R
\name{add_node_factor}
\alias{add_node_factor}
\title{Terminal Node Indicator}
\usage{
add_node_factor(object, newdata, important = FALSE, ...)
}
\arguments{
\item{object}{An object that inherits from class \code{rpart}.}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{important}{Logical indicating whether to include only important
variables. Default is \code{FALSE}.}

\item{...}{Additional optional arguments. At present, no optional arguments
  are used.}
}
\value{
A data frame containing the original data with an additional factor
  variable containing terminal node assignments.
}
\description{
Augments the supplied data with an additional factor variable describing
terminal node assignment.
}

