% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num.R
\name{num_unique}
\alias{num_unique}
\title{Number of Unique Values}
\usage{
num_unique(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
\code{num_unique} returns the number of unique values in \code{x}.
}

