% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num.R
\name{numeric_columns}
\alias{numeric_columns}
\title{Extract Numeric Columns}
\usage{
numeric_columns(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
\code{numeric_columns} returns a vector of names corresponding to the numeric
columns of \code{x}.
}

