% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/n.R
\name{nunique}
\alias{nunique}
\alias{nunique.data.frame}
\alias{nunique.default}
\alias{nunique.matrix}
\title{The Number of Unique Elements}
\usage{
nunique(x)

\method{nunique}{default}(x)

\method{nunique}{matrix}(x)

\method{nunique}{data.frame}(x)

nunique(x)
}
\arguments{
\item{x}{A vector, matrix, or data frame.}

\item{x}{A vector.}
}
\description{
Return the number of unique values which its argument has.

\code{num_unique} returns the number of unique values in \code{x}.
}

