% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/magic.R
\name{odd_magic_square}
\alias{odd_magic_square}
\title{Odd Magic Square}
\usage{
odd_magic_square(n)
}
\arguments{
\item{n}{An odd integer}
}
\description{
Construct a magic square of odd order using de La Loubere’s method.
}
\examples{
oms <- odd_magic_square(9)
rowSums(oms)
colSums(oms)
sum(diag(oms))
}

