% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ones.R
\name{ones}
\alias{ones}
\alias{zeros}
\title{Create an array of ones, zeros}
\usage{
ones(...)

zeros(...)
}
\arguments{
\item{\dots}{A list of the number of ones/zeros along each dimension.
Can also be an integer vector.}
}
\value{
Returns an \code{array} of the dimension specified.
}
\description{
Simple functions for creating an array of ones or zeros. Simply a wrapper
for \code{array}.
Built for convinienece and not speed.
}
\examples{
ones(10)
ones(10,5)
ones(c(2,3,3))
ones(3,3,3, dimnames=list(LETTERS[1:3], letters[1:3], LETTERS[4:6]))

zeros(10, 12)
zeros(c(5, 3, 2))
zeros(3, 4, dimnames=list(LETTERS[1:3], letters[9:12]))
}
\seealso{
\code{\link{array}}
}

