% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{orig2t_inner}
\alias{orig2t_inner}
\title{orig2t_inner}
\usage{
orig2t_inner(orig, original_time.periods)
}
\arguments{
\item{orig}{a single original time period to convert to new time period}

\item{original_time.periods}{vector containing all original time periods.}
}
\description{
A helper function to switch from original t values to
"new" t values (which are just time periods going from 1 to total
number of available periods).
}
\keyword{internal}
