% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{panel2cs}
\alias{panel2cs}
\title{Panel Data to Repeated Cross Sections}
\usage{
panel2cs(data, timevars, idname, tname)
}
\arguments{
\item{data}{data.frame used in function}

\item{timevars}{vector of names of variables to keep}

\item{idname}{unique id}

\item{tname}{time period name}
}
\value{
data.frame
}
\description{
panel2cs takes a 2 period dataset and turns it
 into a cross sectional dataset.  The data includes the
 change in time varying variables between the
 time periods.  The default functionality
 is to keep all the variables from period 1
 and add all the variables listed by name in timevars
 from period 2 to those.
}
\examples{
id <- rep(seq(1, 50), 2)
t <- rep(seq(1, 2), 50)
y <- rnorm(100)
dta <- data.frame(id = id, t = t, y = y)
out <- panel2cs(dta, timevars = "y", idname = "id", tname = "t")
nrow(out)

}
