% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partial.R
\name{partial_values}
\alias{partial_values}
\title{Partial Dependence Values}
\usage{
partial_values(object, xname, xvalues, newdata, ...)
}
\arguments{
\item{object}{An appropriate R object.}

\item{xname}{Character string specifying the predictor variable of interest.}

\item{xvalues}{Vector of values of \code{xname} for which partial dependence
values are sought.}

\item{newdata}{Data frame containing data for all variables of interest in
the model.}

\item{...}{Additional optional arguments to be passed onto \code{predict}.}
}
\description{
Compute partial dependence values from an appropriate R object.
}
\note{
Only used for regression (i.e.,continuous response values). This may
  change in the future. Also, this function only calculates partial
  dependence for one independent variabe at a time. Bivariate partial
  dependence will be available in a future release.
}

