% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pcor.R
\name{pcor}
\alias{pcor}
\title{Compute first order partial correlations}
\usage{
pcor(S, k)
}
\arguments{
\item{S}{A numeric correlation matrix}

\item{k}{An integer. The variable to condition on.}
}
\value{
A matrix of the same size as \code{X} with \code{NaN}s in the
  \code{k} row and column.
}
\description{
This function computes the first order partial correlation given the
\code{k}'th variable.
}
\examples{
X <- replicate(4, rnorm(10))
dimnames(X) <- list(paste0("obs", 1:nrow(X)), paste0("var", 1:ncol(X)))
S <- cor(X)
pcor(S, 2)
}
\author{
Anders Ellernsta Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link{cor}}
}

