% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.multi.R
\name{phi.coefficient}
\alias{phi.coefficient}
\title{Phi coefficient}
\usage{
phi.coefficient(x)
}
\arguments{
\item{x}{A 2-by-2 contigency table.}
}
\value{
The phi coefficient for the table.
}
\description{
Computes the phi coefficient measure of association for 2-by-2 contigency 
tables.
}
