% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.multi.R
\name{phi.multi}
\alias{phi.multi}
\title{Phi Coefficient for Multi Label Data}
\usage{
phi.multi(x, y)
}
\arguments{
\item{x}{Rater 1.}

\item{y}{Rater 2.}
}
\value{
A list with the following components: \code{avg.phi} and \code{phi}.
}
\description{
Computes an averaged phi coefficient measure of association for multi label
data.
}

