% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R
\name{plot.battery_target}
\alias{plot.battery_target}
\title{Variable Importance Heatmap}
\usage{
\method{plot}{battery_target}(x, ...)
}
\arguments{
\item{x}{An object that inherits from class \code{"battery_target"}.}

\item{...}{Additional optional arguments to be passed onto \code{d3heatmap}.}
}
\description{
Produces a heatmap of all the pairwise variable importance scores.
}

