% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotChromosome}
\alias{plotChromosome}
\title{Plot a chromosome}
\usage{
plotChromosome(chromosome = "chr1", xleft = 0, ybot = 0, xright = 1,
  ytop = 1, taper = 0.2, build = "hg19", ...)
}
\arguments{
\item{chromosome}{A character giving the chromosome to plot. Should be one
of \code{"chr1"}, ..., \code{"chr22"}, \code{"chrX"}, or \code{"chrY"}.}

\item{xleft}{A scalar giving the left x position.}

\item{ybot}{A scalar giving the bottom y postion.}

\item{xright}{A scalar giving the right x postion.}

\item{ytop}{A scalar giving the top y postion.}

\item{taper}{A numeric givning the amount of tapering near the ends of the
chromosome and the centromere.}

\item{build}{UCSC genome build. Supported builds are "hg18" and "hg19".}

\item{\dots}{Arguments passed to polygon.}
}
\description{
\code{plotChromosome} draws a chromosome with an interface akin to
\code{rect} with the given coordinates.
}
\note{
Some of the code is borrowed from \code{plotIdiogram} in the \code{SNPchip}
  package by Robert Scharpf and Jason Ting.
}
\examples{
dat <- cumsum(rnorm(100))/sqrt(seq_len(100))
  plot(dat, ylab = "Something", type = "h", ylim = c(-3, 3),
       col = ifelse(dat > 0, "red", "blue"), lwd = 2)

  plotChromosome("chr10", xleft = 0, ybot = -3, xright = 100, ytop = -2.8)
}
\author{
Anders Ellern Bilgrau \cr
  Modified code by Robert Scharpf and Jason Ting
}
\seealso{
\code{\link[SNPchip]{plotIdiogram}} \cr
  \code{\link[oligoClasses]{getSequenceLengths}}
}

