% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotColourKey.R
\name{plotColourKey}
\alias{plotColourKey}
\title{Plot a colour key}
\usage{
plotColourKey(breaks, col, add = FALSE, ...)
}
\arguments{
\item{breaks}{A numeric vector of length \eqn{n}
giving the breakpoints of the key.}

\item{col}{A character vector of the colours used of length \eqn{n - 1}.}

\item{add}{Should the key be added to the current plot device?}

\item{\dots}{Arguments passed to \code{\link{rect}}.}
}
\description{
Easy plotting of a colour key.
}
\examples{
brks <- seq(0, 1, length.out = 8)
cls <- rainbow(7)
plotColourKey(brks, cls)
axis(1)
axis(2)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}

