% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotColourMat.R
\name{plotColourMat}
\alias{plotColourMat}
\title{Plot a matrix of colors}
\usage{
plotColourMat(x, add = FALSE, ...)
}
\arguments{
\item{x}{A \code{character} \code{matrix} of colour names.}

\item{add}{\code{logical}. Should the matrix be added to the current plot
device.}

\item{\dots}{Arguments passed to \code{\link{rect}}.}
}
\value{
Invisibly returns the x and y coordinates and the corresponding colours.
}
\description{
Plot a \code{character} matrix giving the colours to plot.
}
\details{
The colours are plotted as squares as given by \code{x} on the grid
  of integers. The entry given in \code{x[i,j]} is plotted as a rectangle
  at coordinate (\code{i}, \code{nrow(x)-j}).
}
\examples{
plotColourMat()  # Plots a random 6 by 10 colour matrix.

x <- matrix(sample(colors(), 20), ncol = 5)
plotColourMat(x)

x <- matrix(rainbow(100), 1, 100)
plotColourMat(x)
dim(x) <- c(10, 10)
plotColourMat(x)

y <- matrix(sample(colors(), 25), ncol = 5)
plotColourMat(y, add = TRUE)
axis(1)
axis(2)
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

