% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotHierarchicalEdgeBundles}
\alias{plotHierarchicalEdgeBundles}
\title{Hierarchical Edge Bundling}
\usage{
plotHierarchicalEdgeBundles(phylo, graph, beta = 0.5, use.mrca = TRUE,
  simplify = FALSE, ..., args.lines = list(), debug = FALSE)
}
\arguments{
\item{phylo}{A `phylo` object.}

\item{graph}{A \code{igraph} object.}

\item{beta}{The amount of bundling.}

\item{use.mrca}{Should the only the most recent common ancestor or the
full shortest path be used?}

\item{simplify}{Simplify the paths by taking the convex hull.}

\item{...}{Arguments passed to \code{\link[ape]{plot.phylo}}.}

\item{args.lines}{A list of arguments passed to \code{lines}.}

\item{debug}{Plot some extra info.}
}
\value{
Plots to a new device.
}
\description{
Visualization of networks using Hierarchical Edge Bundles.
}
\examples{
if (require("igraph")) {
  graph <- erdos.renyi.game(n = 200, p = 0.05)
  wt <- walktrap.community(graph, modularity=TRUE)
  phylo <- asPhylo(wt)
  plotHierarchicalEdgeBundles(phylo,
                              graph,
                              beta = 0.2,
                              use.mrca = FALSE,
                              type = "unrooted",
                              use.edge.length = TRUE,
                              simplify = FALSE,
                              lab4ut = "axial",
                              args.lines = list(col = "#FF000060"))
}
}
\author{
Anders Ellern Bilgrau
}
\references{
Holten, Danny. "Hierarchical edge bundles: Visualization of adjacency
  relations in hierarchical data." Visualization and Computer Graphics, IEEE
  Transactions on 12.5 (2006): 741-748.
}
\seealso{
See \code{\link[ape]{plot.phylo}}.
}

