% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotModuleGraph}
\alias{plotModuleGraph}
\title{Plot the graphical representation of a weigted adjacency amatrix}
\usage{
plotModuleGraph(amat, diff.exprs = NULL, label = rownames(amat),
  vcol = "Black", layout = layout.circle, ...)
}
\arguments{
\item{amat}{A numeric adjacency amatrix.}

\item{diff.exprs}{A numeric vector of lencth \code{ncol(amat)} giving the
importance (e.g.\ evidence for differential expression) of the node/gene.}

\item{label}{The labels of the nodes.}

\item{vcol}{The colors of the nodes. Default is black.}

\item{\dots}{Aguments passed to \code{plot.igraph}.}

\item{layout}{A function providing the layout of the graph.
See \link[igraph]{layout}.}
}
\description{
A function of for plotting the graphical representation of a weigted
weigted adjacencey amatrix. Build for convenience and not for speed.
}
\examples{
# Construct a random adjacency matrix
amat <- replicate(10, runif(10))
amat[upper.tri(amat)] <- amat[lower.tri(amat)]

plotModuleGraph(amat,
                label = "",
                vcol = 1:10,
                diff.exprs = 3 + 3*runif(10),
                )
}
\author{
Anders Ellern Bilgrau <abilgrau (at) amath.aau.dk>
}

