% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModuleImage.R
\name{plotModuleImage}
\alias{plotModuleImage}
\title{Plot the heatmap of a adjacency matrix}
\usage{
plotModuleImage(amat, print.xlab = TRUE, print.ylab = TRUE,
  labels = rownames(amat), srt = 45, xlab.at = par("usr")[3] - 0.25,
  ylab.at = par("usr")[1] - 0.25, ...)
}
\arguments{
\item{amat}{An numeric adjacency matrix with values between 0 and 1.}

\item{print.xlab}{logical. Should the labels be shown at the x-axis?}

\item{print.ylab}{logical. Should the labels be shown at the y-axis?}

\item{labels}{character vector of labels to be printed at the x-axis and
y-axis.}

\item{srt}{numeric. Rotation of the labels. See \code{\link{par}}.}

\item{xlab.at}{The position of the labels in of the x-axis.}

\item{ylab.at}{The position of the labels in of the y-axis.}

\item{\dots}{Arguments passed to \code{\link{image}}.}
}
\description{
Plot the heatmap of a adjacency matrix
}
\examples{
# Construct a random adjacency matrix
amat <- replicate(10, runif(10))
amat[upper.tri(amat)] <- amat[lower.tri(amat)]
diag(amat) <- NA
rownames(amat) <- colnames(amat) <- paste0("Gene", 1:10)

# Plot the graph
plotModuleImage(amat, srt = 30)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
