% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\alias{pop.default}
\alias{pop.list}
\title{Remove Items}
\usage{
pop(x, i = length(x))

\method{pop}{default}(x, i = length(x))

\method{pop}{list}(x, i = length(x))
}
\arguments{
\item{x}{An appropriate R object.}

\item{i}{Integer position of the element to be removed in \code{x}. Defaults
to removing the last element of \code{x}.}
}
\value{
Returns the object \code{x} with the i-th element removed.
}
\description{
\code{pop} is a generic function that will remove the item at the given 
position.
}
\examples{
x <- 1:3
y <- letters[1:3]
z <- as.list(x)
names(z) <- y
pop(x)
pop(y, 2)  # remove the second element
pop(y, -2)  # remove everythng except the scond element
for (i in 1:3) print(z <- pop(z))
}

