% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.R
\name{pwr.logRank}
\alias{pwr.logRank}
\title{Sample size calculations for the log-rank test}
\usage{
pwr.logRank(S.trt, S.ctrl, sig.level = 0.05, power = 0.8,
  alternative = c("two.sided", "less", "greater"),
  method = c("Freedman"))
}
\arguments{
\item{S.trt}{Numeric specifying the hazard rate for the treatment group.}

\item{S.ctrl}{Numeric specifying the hazard rate for the control group.}

\item{sig.level}{Significance level (Type I error probability).}

\item{power}{Power of test (1 minus Type II error probability).}

\item{alternative}{Character string specifying the form of the alternative 
hypothesis. Must be one of \code{"two.sided"} (default), \code{"greater"} 
or \code{"less"}.}

\item{method}{Character string specifying the method to use in calculating
the required sample size. (Currently ignored.)}
}
\value{
An estimate of the required sample size.
}
\description{
Calculates the required sample size for a log-rank test using Freedman's
method.
}
\details{
PASS Sample Size Software, Chapter 700:
The power calculations used here assume an underlying exponential 
distribution. However, we are rarely in a position to assume exponential 
survival times in an actual clinical trial. How do we justify the exponential 
survival time assumption? First, the logrank test and the test derived using 
the exponential distribution have nearly the same power when the data are in 
fact exponentially distributed. Second, under the proportional hazards model 
(which is assumed by the logrank test), the survival distribution can be 
transformed to be exponential and the logrank test remains the same under 
monotonic transformations.
}
\note{
Returns \code{Inf} when \code{S.trt == S.ctrl}.
}
\examples{
pwr.logRank(0.3, 0.4)
}
