\name{r.colors}

\alias{r.colors}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Pie charts of all R character colors
}
\description{
Creates a pdf file with pie charts of all the 657 basic character colors of R
}
\usage{
r.colors(file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{
the directory in which the pdf file will be created
}

}
\details{
Define the directory in which the file should saved by writing \code{file="C:/temp"} for example.
If file is not defined, it will be saved in \code{"C:/"} on windows and in \code{"home"} on Mac.
}
\value{
None
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
r.colors()
}

