% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{rhs}
\alias{rhs}
\title{Right-hand Side of Formula}
\usage{
rhs(formla)
}
\arguments{
\item{formla}{a formula}
}
\value{
a one sided formula
}
\description{
Take a formula and return the right hand side
 of the formula
}
\examples{
ff <- yvar ~ x1 + x2
rhs(ff)

}
