% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{rhs.vars}
\alias{rhs.vars}
\title{rhs.vars}
\usage{
rhs.vars(formla)
}
\arguments{
\item{formla}{a formula}
}
\value{
vector of variable names
}
\description{
Take a formula and return a vector of the variables
 on the right hand side
}
\examples{
ff <- yvar ~ x1 + x2
rhs.vars(ff)

ff <- y ~ x1 + I(x1^2)
rhs.vars(ff)

}
