% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{rhs_vars}
\alias{rhs_vars}
\title{Right-hand Side Variables}
\usage{
rhs_vars(formula)
}
\arguments{
\item{formula}{a formula}
}
\value{
vector of variable names
}
\description{
Take a formula and return a vector of the variables
 on the right hand side
}
\examples{
ff <- yvar ~ x1 + x2
rhs_vars(ff)

ff <- y ~ x1 + I(x1^2)
rhs_vars(ff)

}
