% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_vec.R
\name{row_vec}
\alias{row_vec}
\alias{row_vec.default}
\alias{row_vec.matrix}
\alias{row_vec.data.frame}
\title{Row Vectors}
\usage{
row_vec(..., use.names = TRUE)

\method{row_vec}{default}(..., use.names = TRUE)

\method{row_vec}{matrix}(..., use.names = TRUE)

\method{row_vec}{data.frame}(..., use.names = TRUE)
}
\arguments{
\item{...}{The elements of the row vector.}

\item{use.names}{Logical indicating whether or not names should be preserved.}
}
\value{
A row vector of dimension \code{1 x length(x)}.
}
\description{
Constructs a row vector.
}
\examples{
row_vec(1:5)
row_vec(1, 2, 3, 4, 5)
row_vec(t(row_vec(1:5)))
row_vec(col_vec(1:5))
row_vec(pi, exp(1), sin(0))
row_vec(rnorm(3))
row_vec("a" = 1, "b" = 2)  # names are preserved by default
}
