% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rwishart}
\alias{rinvwishart}
\alias{rwishart}
\title{Wishart and inverse-Wishart distributions}
\usage{
rwishart(sigma, nu)

rinvwishart(Psi, nu)
}
\arguments{
\item{sigma}{A square symmetric postive-definite matrix.}

\item{nu}{An integer giving the degrees of freedom.}

\item{Psi}{A square symmetric postive-definite matrix.}
}
\value{
\code{rwishart} returns a matrix drawn from the Wishart distribution.
}
\description{
Generate data from a Wishart and inverse-Wishart distribution.
}
\seealso{
\code{\link{rWishart}}
}

