% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/saltedHash.R
\name{saltedHash}
\alias{saltedHash}
\title{Salted hashing function}
\usage{
saltedHash(salt, x, algo = "sha256", ...)
}
\arguments{
\item{salt}{A character containing the salt}

\item{x}{A character vector of the strings to be hashed.}

\item{algo}{The hashing algorithm used. See \code{\link[digest]{digest}} for
possible values.}

\item{\dots}{Additional parameters passed to \code{\link[digest]{digest}}.}
}
\value{
A character vector
}
\description{
This function adds salt (stochastic or deterministic) to a character
of strings prior to a hashing of the complete string.
}
\examples{
x <- c("AVerySecretString", "AnotherSecretSting")
saltedHash("SomeSalt", x = x, algo = "md5")
#digest(paste0("SomeSalt", "AVerySecretString"), algo = "md5")
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link[digest]{digest}}
}

