% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.remove.R
\name{save.remove}
\alias{save.remove}
\title{Remove object from .RData file}
\usage{
save.remove(rmv, file, ...)
}
\arguments{
\item{rmv}{A \code{character} vector giving the names of the object to 
remove from the specified file.}

\item{file}{A \code{character} of length one specifying the \code{.RData} 
file to remove objects from.}

\item{\dots}{Arguments passed to \code{save}.}
}
\description{
Function to remove one or more objects from an \code{.RData} file.
}
\examples{
file <- paste(tempfile(), ".RData", sep = "")
a <- 1
b <- list(1:2)
c <- lm(rnorm(100) ~ 1)
save(a, b, c, file = file)
rm(a, b, c)
save.remove(c("a", "b"), file = file)
lst <- load(file)
print(lst)  # Only "c" remains
}
\author{
Anders E Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
