% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{set_seed}
\alias{set_seed}
\title{Specify seeds}
\usage{
set_seed(string)
}
\arguments{
\item{string}{A character string.}
}
\value{
Returns NULL, invisibly.
}
\description{
Specify seeds using character starins.
}
\examples{
set_seed("Some reproducible random numbers")
rnorm(3)  # should give: 1.4243634 -0.8759381 -1.6210951
}
