% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str2int.R
\name{setSeed}
\alias{setSeed}
\title{Specify Seeds}
\usage{
setSeed(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
Returns NULL, invisibly.
}
\description{
Specify seeds using character vector.
}
\examples{
setSeed("Some random numbers")
rnorm(3)  # should give: 1.7500983 -0.1093635 -0.9958618
}

