% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_ris}
\alias{sim_ris}
\title{Simulate random intercept and slope data}
\usage{
sim_ris(m = 30, n = 10, mu = c(4.8, 3), sigma = c(1.3, 0.7, 0.7))
}
\arguments{
\item{m}{Integer specifying the number of subjects.}

\item{n}{Integer specifying the number of observations per subject.}

\item{mu}{Numeric vector of length two giving the means of the random 
intercept and slope, respectively.}

\item{sigma}{Numeric vector of length three giving the standard deviations of 
the random intercept, slope, and error term, respectively.}
}
\description{
Simulate data from a (balanced) random intercept and slope model.
}
\examples{
set.seed(101)  # for reproducibility
df <- sim_ris(m = 100, n = 15)
}
