% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{split_aachange}
\alias{split_aachange}
\title{Split protein sequence variant column into individual columns
Divides column containing protein sequence variants into unique reference,
position, and variant columns.}
\usage{
split_aachange(df, column = "aachange")
}
\arguments{
\item{df}{dataframe with protein sequence variant column}

\item{column}{name of the column to split in df}
}
\value{
updated df with reference_aa, position_aa, and alternate_aa columns
}
\description{
Split protein sequence variant column into individual columns
Divides column containing protein sequence variants into unique reference,
position, and variant columns.
}
