% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R
\name{splitters}
\alias{splitters}
\title{Extract Splitting Variables}
\usage{
splitters(x)
}
\arguments{
\item{x}{An object that inherits from class \code{"rpart"}.}
}
\description{
\code{splitters} returns a vector of variable names used to partition
the data.
}
\keyword{internal}

