% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statmode.R
\name{statmode}
\alias{statmode}
\title{Statistical Mode}
\usage{
statmode(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numbers or characters.}

\item{na.rm}{Logical indicating whether \code{NA} values should be stripped
before the computation prceeds.}
}
\description{
Compute the mode (i.e., most frequent element).
}

