% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str2int}
\alias{str2int}
\title{Convert characters to integers}
\usage{
str2int(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
An integer (not necessarily unique).
}
\description{
Convert one-dimensional character objects to integers.
}
\examples{
# Should all produce the same integer (currently results in 124)
str2int("Simulation 1")
str2int("simulation 1")
str2int("SIMULATION 1")
str2int("simulation.1")
str2int("simulation..1")
str2int("simulation-1")
str2int("simulation - 1")
str2int("simulation---1")
str2int("simulation1")

# Potential issues
str2int("abc")  # does not provide unique integers
str2int("cba")
}
