% Generated by roxygen2 (4.0.2): do not edit by hand
\name{straightenedBezier}
\alias{straightenedBezier}
\title{Straightened Bezier curves}
\usage{
straightenedBezier(x, y, beta = 0.5, evaluation = 100)
}
\arguments{
\item{x}{Some x-values}

\item{y}{Some y-values (optional)}

\item{beta}{The amount of straightening.}

\item{evaluation}{The number of evaluations of the Bezier curve.}
}
\value{
Returns the
}
\description{
Straightened Bezier curves
}
\author{
Anders Ellern Bilgrau
}
\keyword{internal}

