% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R
\name{summarize_classification_tree}
\alias{summarize_classification_tree}
\title{Summarize Classification Trees}
\usage{
summarize_classification_tree(tree)
}
\description{
\code{summarize_classification_tree} returns a basic summary of a
classification tree.
}
\keyword{internal}

