% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num.R
\name{summarize_columns}
\alias{summarize_columns}
\title{Summarize Columns}
\usage{
summarize_columns(x)
}
\arguments{
\item{x}{A data frame or an R object that can be coerced into one.}
}
\description{
\code{summarize_columns} produces a data frame summarizing the columns of
\code{x}.
}
\examples{
d <- data.frame(w = c(NA, NA, exp(1:2), NA),
                x = 1:5,
                y = c("a", "b", "a", "a", NA),
                z = c(T, T, F, NA, NA),
                stringsAsFactors = FALSE)
summarize_columns(d)
}

