% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data}
\usage{
summarize_data(x, ...)
}
\arguments{
\item{x}{A data frame or an R object that can be coerced into one.}

\item{...}{Additional optional arguments.}
}
\description{
Summarize a data frame.
}

