% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R
\name{summarize_regression_tree}
\alias{summarize_regression_tree}
\title{Summarize Regression Trees}
\usage{
summarize_regression_tree(tree, na.rm = FALSE, formula = "corr")
}
\description{
\code{summarize_regression_tree} returns a basic summary of a regression
tree.
}
\keyword{internal}

