% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R
\name{summarize_tree}
\alias{summarize_tree}
\title{Summarize Decision Trees}
\usage{
summarize_tree(object, na.rm = FALSE, formula = "corr")
}
\arguments{
\item{object}{An object that inherits from class \code{"rpart"}.}

\item{na.rm}{Logical indicating whether or not \code{NA}s should be removed.
Default is \code{FALSE}.}

\item{formula}{Character string specifying which formula to use in
calculating R-squared. Default is \code{"corr"}, meaning use the squared
correlation between the observed and predicted values.  The only other
option is \code{"traditional"}. See Kvalseth (1985) for a summary of the
different approaches.}
}
\description{
\code{summarize_tree} returns a basic summary of a decision tree.
}
\keyword{internal}

