% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeData.R
\name{summarizeData}
\alias{summarizeData}
\title{Summarize Data}
\usage{
summarizeData(x, ...)
}
\arguments{
\item{x}{A data frame or an R object that can be coerced into one.}

\item{...}{Additional optional arguments.}
}
\description{
Summarize a data frame.
}

