% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{t2orig}
\alias{t2orig}
\title{t2orig}
\usage{
t2orig(t, original_time.periods)
}
\arguments{
\item{t}{a vector of time periods to convert back to original time
periods.}

\item{original_time.periods}{vector containing all original time periods.}
}
\value{
original time period converted from new time period
}
\description{
A helper function to switch from "new" t values to
original t values.  This allows for periods not being exactly spaced
apart by 1.
}
\examples{
original_time.periods <- c(2001, 2003, 2005, 2007)
t2orig(1:4, original_time.periods) ## returns c(2001, 2003, 2005, 2007)

}
