% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{t2orig_inner}
\alias{t2orig_inner}
\title{t2orig_inner}
\usage{
t2orig_inner(t, original_time.periods)
}
\arguments{
\item{t}{a single time period to convert back to original time}

\item{original_time.periods}{vector containing all original time periods.}
}
\description{
A helper function to switch from "new" t values to
original t values for a single t.
}
