% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timing.R
\name{timing}
\alias{timing}
\title{Easy function and evalutation timing}
\usage{
timing(expr)
}
\arguments{
\item{expr}{A valid R expression or function call.}
}
\value{
Returns the result of the evaluated \code{expr} with an attribute
  called \code{"timing"} containing the evaulation time.
}
\description{
Times a function or evaulation and appends the timing results as an
attribute to the object.
}
\examples{
test <- function() mean(rnorm(1e6))
a <- timing(test())
print(a)

b <- timing(mean(rnorm(1e6)))
print(b)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}

