% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{toformula}
\alias{toformula}
\title{Variable Names to Formula}
\usage{
toformula(yname, xnames)
}
\arguments{
\item{yname}{the name of the y variable}

\item{xnames}{vector of names for x variables}
}
\value{
a formula
}
\description{
take a name for a y variable and a vector of names
 for x variables and turn them into a formula
}
\examples{
toformula("yvar", c("x1", "x2"))

## should return yvar ~ 1
toformula("yvar", rhs_vars(~1))

}
