% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{TorF}
\alias{TorF}
\title{TorF}
\usage{
TorF(cond, use_isTRUE = FALSE)
}
\arguments{
\item{cond}{a vector of conditions to check}

\item{use_isTRUE}{whether or not to use a vectorized version
of isTRUE.  This is generally slower but covers more cases.}
}
\value{
logical vector
}
\description{
A function to replace NA's with FALSE in vector of logicals
}
\examples{
TorF(c(TRUE, NA, FALSE)) ## NA becomes FALSE

}
