% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_utils.R
\name{transpose.data}
\alias{transpose.data}
\title{Swap axes of the dataframe. Patients are now rows and other data columns}
\usage{
transpose.data(df)
}
\arguments{
\item{df}{Dataframe with patient info as columns and other data as rows}
}
\value{
Dataframe where each row is a patient and each column is some type of TCGA variable
}
\description{
Swap axes of the dataframe. Patients are now rows and other data columns
}
