% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_summary}
\alias{tree_summary}
\title{Battery Target Tree Summary}
\usage{
tree_summary(x, scale = TRUE, ...)
}
\arguments{
\item{x}{A list with three components: \code{tree}, \code{response}, and
\code{predictors}.}

\item{scale}{Logical indicating wether or not variable importance scores
should be scaled to sum to one. Default is \code{TRUE}.}

\item{...}{Additional optional arguments. None are currently used.}
}
\description{
Summarize a single battery target tree.
}

