% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{weighted.checkfun}
\alias{weighted.checkfun}
\title{Weighted Check Function}
\usage{
weighted.checkfun(q, cvec, tau, weights)
}
\arguments{
\item{q}{the value to check}

\item{cvec}{vector of data to compute quantiles for}

\item{tau}{between 0 and 1, ex. .5 implies get the median}

\item{weights}{the weights, weighted.checkfun normalizes the weights
to sum to 1.}
}
\value{
numeric
}
\description{
Weights the check function
}
