% Generated by roxygen2 (4.0.1): do not edit by hand
\name{welch.t.tests}
\alias{welch.t.tests}
\title{Perform rowwise Welch t-tests.}
\usage{
welch.t.tests(x, y)
}
\arguments{
\item{x}{A numeric matrix of the values of the first group. Usually the
expression matrix of the first group.}

\item{y}{A numeric matrix of the values in the second group with the same
number of rows as \code{x}.}
}
\value{
A \code{data.frame} with the same number of rows as x containing the
  means, variances, pooled stadard
  deviations, estimated difference, the standard error of the difference,
  the t statistic, the degrees of freedom, and the p-values.
}
\description{
This functions performs a Welch two-sample t-test for each row in two
matrices with the same height in a fast manner.
}
\details{
A Welch t-test is perform for each row in \code{x} and \code{y}.
  The i'th row of \code{x} is tested against the i'th row of \code{y}
  and the result is seen in the i'th row of the output.
}
\examples{
x <- replicate(10,   rnorm(20, 0, 1))
y <- replicate(11, c(rnorm(10, 1, 1), rnorm(10, 0, 1)))

welch.t.tests(x, y)

# To perform a single t test
t.test(x[1,], y[1, ])
welch.t.tests(x[1, , drop = FALSE], y[1, , drop = FALSE])
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

