% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which_numeric}
\alias{which_numeric}
\alias{which_factor}
\alias{which_character}
\alias{which_integer}
\title{Where are certain types of columns?}
\usage{
which_numeric(x, names = TRUE)

which_factor(x, names = TRUE)

which_character(x, names = TRUE)

which_integer(x, names = TRUE)
}
\arguments{
\item{x}{A data frame (ideally), but list-like objects should work too.}

\item{names}{Logical indicating whether to return names (\code{TRUE}) or 
indexes \code{FALSE}. Default is \code{TRUE}.}
}
\description{
Determines the names or locations (i.e., index) of particular types of 
columns.
}
\examples{
which_numeric(iris)
which_factor(iris)
d <- data.frame(x = 1:3, y = c("a", "b", "c"), stringsAsFactors = FALSE)
which_integer(d)
which_numeric(d)
which_character(d)
}
