% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeForNetsTensor.R
\name{writeForNetsTensor}
\alias{writeForNetsTensor}
\title{Format a list of networks for the "NetsTensor" software.}
\usage{
writeForNetsTensor(x.list, out.dir = "netsdata", verbose = TRUE)
}
\arguments{
\item{x.list}{A list of matrices of the same size}

\item{out.dir}{character. The path to the output dir where the files are to
be written.}

\item{verbose}{logical. Should the progress be outputted?}
}
\value{
Returns invisibly a \code{list} of length 4 with the content of the
  written files. Writes the formatted networks to
  the disk in the directory specified by \code{out.dir}.
}
\description{
This function formats the networks for the "NetTensor" software by Li et. al.
(see references below) and writes the result to the harddrive.
}
\note{
The software is available at the website
  \url{http://zhoulab.usc.edu/tensor/}.
}
\references{
Li, W., Liu, C. C., Zhang, T., Li, H., Waterman, M. S., & Zhou, X. J.
  (2011). Integrative analysis of many weighted co-expression networks using
  tensor computation. PLoS computational biology, 7(6), e1001106.
}

