% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lda.fit}
\alias{lda.fit}
\alias{qda.fit}
\alias{qda2lda}
\title{Estimate parameters in discriminant analysis}
\usage{
lda.fit(classes, vars)

qda.fit(classes, vars)

qda2lda(myqda)
}
\arguments{
\item{classes}{A factor giving the classes of the rows of \code{var}.}

\item{vars}{A numeric matrix giving the observed quantities. Rows correspond
to observations and columns to variables.}

\item{myqda}{The output of \code{lda.fit}.}
}
\value{
Returns a \code{list} with entries:
  \item{counts}{The number of observations in each class.}
  \item{means}{The estimated means for each class.}
  \item{sigmas}{The estimated covariance matrix for each class.}
}
\description{
Functions to estimate the parameters in linear and quadratic
discriminant analysis (LDA, QDA, resp.). \code{qda2lda} provides
coversion from QDA to LDA.
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

