% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discriminant_analysis.R
\name{lda.predict}
\alias{lda.predict}
\alias{qda.predict}
\title{Classify according to a discriminant analysis fit}
\usage{
lda.predict(fit, newdata)

qda.predict(fit, newdata)
}
\arguments{
\item{fit}{A \code{list} as in the output of \code{qda.fit} or
\code{lda.fit}.}

\item{newdata}{A numeric matrix giving the observed quantities. Rows correspond
to observations and columns to variables.}
}
\value{
Returns a \code{list} with entries:
  \item{class}{A vector giving the predicted classes of the observations.}
  \item{posterior}{A matrix of posterior probabilities of belonging to the
    classes.}
}
\description{
Functions to classify new data on the basis of a LDA/QDA fit.
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}

