% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addCarriageReturn.R
\name{addCarriageReturn}
\alias{addCarriageReturn}
\title{Add Carriage Return}
\usage{
addCarriageReturn(file.in, file.out = file.in)
}
\arguments{
\item{file.in}{The path and name of the input file. A character of length 1.}

\item{file.out}{The path and name of the output file. A character of
length 1. Default is to overwrite.}
}
\value{
Nothing is returned. Creates (or overwrites if file.out is not
  given) a file with added carriage returns.
}
\description{
Unix based systems use only the line feed (\code{LF}, \code{"\\n"}) character
for line breaks whereas Windows both line feed and carriage return
(\code{CR}, \code{"\\r"}). This function adds the carriage return in order
for files to be read properly on windows.
}
\note{
Useful on Unix systems to make Windows readable \code{.txt} and other
  plain-text files.
}
\examples{
tmp.dir <- tempdir()
file.in <- file.path(tmp.dir, "NewFile.txt")
file.out <- file.path(tmp.dir, "NewFileForWindows.txt")

# Write the file
cat("This is some text on.\\n Hello, newline.", file = file.in)

# Add carriage return if on a unix system
addCarriageReturn(file.in, file.out)

# If on unix, run to see files:
#system(paste("open", file.in))
#system(paste("open", file.out))
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}

