% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{addCovToFormla}
\alias{addCovToFormla}
\title{addCovToFormla}
\usage{
addCovToFormla(covs, formla)
}
\arguments{
\item{covs}{should be a list of variable names}

\item{formla}{which formula to add covariates to}
}
\value{
formula
}
\description{
\code{addCovFromFormla} addssome covariates to a formula;
  covs should be a list of variable names
}
\examples{
formla <- y ~ x 
addCovToFormla(list("w","z"), formla)

formla <- ~x
addCovToFormla("z", formla)

}
