% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alp}
\alias{alp}
\title{Add transparency to colours}
\usage{
alp(x, alpha)
}
\arguments{
\item{A}{character of colours. Either as names (e.g. \code{"blue"}) or
hexadecimal (\code{"#00FF00"}).}

\item{A}{numeric value with the alpha level. If \code{alpha = 0} is
completely transparent and  \code{alpha = 1} is opaque.}
}
\value{
A character of hexadecimal values with alpha value added.
}
\description{
The function adds transparency to a character of colours names or
hexadecimal values.
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

