% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alp.R
\name{alp}
\alias{alp}
\title{Add transparency to colours}
\usage{
alp(x, alpha)
}
\arguments{
\item{x}{A character of colours. Either as names (e.g. \code{"blue"}) or 
hexadecimal (\code{"#00FF00"}).}

\item{alpha}{A numeric vector of values with the alpha level. 
If \code{alpha = 0} is 
completely transparent and  \code{alpha = 1} is opaque.}
}
\value{
A character of hexadecimal values with alpha value added.
}
\description{
The function adds transparency to a character of colours names or 
hexadecimal values.
}
\examples{
x <- c("Blue", "#402000", NA, "Tomato", "#00FF0250")
alp(x, alpha = 0.5)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
