% Generated by roxygen2 (4.0.2): do not edit by hand
\name{battery}
\alias{battery}
\alias{battery.rpart}
\alias{listBatteryMethods}
\title{Batteries of Runs}
\usage{
battery(object, ...)

\method{battery}{rpart}(object, type = c("priors", "shaving", "target",
  "depth"), newdata, .progress = TRUE, prior.grid, prior.min = 0.05,
  prior.max = 0.95, prior.step = 0.05, ...)

listBatteryMethods()
}
\arguments{
\item{object}{An object that inherits from class \code{rpart} or
\code{randomForest}.}

\item{...}{Additional optional arguments to be passed onto \code{rpart} or
           \code{randomForest}.}

\item{battery}{The type of battery to perform. For the entire list of
               battery methods use \code{batteryMethods("rpart")}.}
}
\value{
A data frame.

A list with two components:
\itemize{
  \item \code{rpart}
  \item \code{randomForest}
}
}
\description{
Tree-based algorithms are characterized by a substantial number of control
settings. Often the optimal values for many parameters cannot be determined
beforehand and require a trial and error experimental approach. In other
cases, it is desirable to try various settings to study their impact on the
resulting models. CART batteries were designed to automate the most
frequently-occurring modeling situations that require multiple collections
of CART runs.

Prints a list of the available battery methods.
}
\examples{
## Spam example

## Load packages
library(caret)
library(rpart)

## Load and partition data
data(spam, package = "kernlab")
set.seed(101)
trainID <- createDataPartition(spam$type, p = 0.7, list = FALSE, times = 1)
trainSample <- spam[trainID, ]
testSample <- spam[-trainID, ]
## CART
cartFit <- rpart(type ~ ., data = trainSample)
## Battery priors
table(trainSample$type)/nrow(trainSample)  # default priors
res <- battery(cartFit, battery = "priors")
res[order(res$error), ]  # order the results
}
\references{
Salford Systems
}

