% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R, R/trees.R
\name{battery_target}
\alias{battery_target}
\title{Battery Target}
\usage{
battery_target(X, cv = FALSE, surrogates = FALSE, competes = TRUE,
  do.trace = FALSE, ...)

battery_target(X, cv = FALSE, surrogates = FALSE, competes = TRUE,
  do.trace = FALSE, ...)
}
\arguments{
\item{X}{A data frame containing the variables of interest.}

\item{cv}{Logical indicating wether or not to build each tree using cross-
validation.}

\item{surrogates}{Logical indicating wether or not surrogate splits should
contribute to the variable importance calculations. Default is
\code{FALSE}.}

\item{competes}{Logical indicating wether or not competing splits should
contribute to the importance calculations. Default is \code{TRUE}.}

\item{do.trace}{Logical indicating wether to print verbose output as battery
target is run. Default is \code{FALSE}.}

\item{...}{Additional optional arguments to be passed on to \code{rpart}.}

\item{X}{A data frame of predictor variables.}
}
\description{
Battery Target gives you powerful means to automatically explore and extract
all mutual dependencies among predictors. By the word "dependencies," we mean
a potentially nonlinear multivariate relationship that goes way beyond the
simplicity of conventional correlations. Furthermore, as a powerful side
effect, this Battery provides general means for missing value imputation,
which is extremely useful to support those modeling engines that do not
directly handle missing values. Helps identify clusters of related variables.

Automatic extraction of mutual dependencies among predictors. Here, the word
"dependencies" refers to a potentially nonlinear multivariate relationship.
}

