% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{blockBootSample}
\alias{blockBootSample}
\title{blockBootSample}
\usage{
blockBootSample(data, idname)
}
\arguments{
\item{data}{data.frame from which you want to bootstrap}

\item{idname}{column in data which contains an individual identifier}
}
\description{
Legacy name for the function `block_boot_sample`,
 please use that function going forward.  This function will
 eventually be deleted
}
\keyword{internal}
