\name{bmisc}
\alias{bmisc}
\docType{package}
\title{Miscellaneous functions}
\description{
This package has different functions that I have accumulated with time. This is the Alpha version.
}

\details{
For pdf version of the help, write \code{vignette("bmisc")}.
}

\format{
\tabular{ll}{
Package: \tab bmisc\cr
Type: \tab Package\cr
Version: \tab 0.2-14\cr
Date: \tab 2011-11-18\cr
License: \tab LGPL >= 3.0\cr
LazyLoad: \tab yes\cr
}      }

\author{
Benoit Bruneau

Maintainer: Benoit Bruneau <benoit.bruneau1@gmail.com>

}



