\name{bmisc}

\alias{bmisc}
\docType{package}
\title{Miscellaneous functions}
\description{
This package has different functions that I have accumulated with time. I am not the author of all of them even though I have modified most of them. This is the Alpha version.
}

\details{
For pdf version of the help, write \code{vignette("bmisc")}.
}

\format{
\tabular{ll}{
Package: \tab bmisc\cr
Type: \tab Package\cr
Version: \tab 0.2-12\cr
Date: \tab 07-07-2011\cr
License: \tab LGPL >= 3.0\cr

}      }

\author{
Benoit Bruneau

Maintainer: Benoit Bruneau <benoit.bruneau1@gmail.com>

}



