% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Bmisc}
\alias{Bmisc}
\alias{Bmisc-package}
\title{My biostatstical miscellaneous functions}
\description{
My frequently used biostatisical miscellaneous R functions and utilities.
Useful for general microarray analysis, gene network analysis, and other
biostatistical applications.
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

