% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{check_staggered}
\alias{check_staggered}
\title{check_staggered}
\usage{
check_staggered(df, idname, treatname)
}
\arguments{
\item{df}{the data.frame used in the function}

\item{idname}{name of column that holds the unit id}

\item{treatname}{name of column with the treatment indicator}
}
\value{
a logical indicating whether treatment is staggered
}
\description{
A function to check if treatment is staggered in a panel data set.
}
\examples{
n <- 50
id <- rep(seq_len(n), each = 4)
t <- rep(1:4, n)
g <- rep(sample(c(0, 2, 3), n, replace = TRUE), each = 4)
treat <- as.integer(t >= g & g > 0)
dta <- data.frame(id = id, t = t, treat = treat)
check_staggered(dta, idname = "id", treatname = "treat")

}
