% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{check_staggered}
\alias{check_staggered}
\title{check_staggered}
\usage{
check_staggered(df, idname, treatname)
}
\arguments{
\item{df}{the data.frame used in the function}

\item{idname}{name of column that holds the unit id}

\item{treatname}{name of column with the treatment indicator}
}
\value{
a logical indicating whether treatment is staggered
}
\description{
A function to check if treatment is staggered in a panel data set.
}
