% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{check_staggered_inner}
\alias{check_staggered_inner}
\title{check_staggered_inner}
\usage{
check_staggered_inner(this_df, treatname)
}
\arguments{
\item{this_df}{a data.frame, for this function it should be specific to
a particular unit}

\item{treatname}{name of column with the treatment indicator}
}
\description{
A helper function to check if treatment is staggered in a panel data set.
}
\keyword{internal}
