% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{combineDfs}
\alias{combineDfs}
\title{combineDfs}
\usage{
combineDfs(y.seq, dflist, pstrat)
}
\arguments{
\item{y.seq}{sequence of possible y values}

\item{dflist}{list of distribution functions to combine}

\item{pstrat}{weights to put on each distribution function}
}
\value{
ecdf
}
\description{
Combines two distribution functions with given weights by pstrat
}
