% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{compare_binary}
\alias{compare_binary}
\title{Compare Variables across Groups}
\usage{
compare_binary(
  x,
  on,
  dta,
  w = rep(1, nrow(dta)),
  report = c("diff", "levels", "both")
)
}
\arguments{
\item{x}{variables to run regression on}

\item{on}{binary variable}

\item{dta}{the data to use}

\item{w}{weights}

\item{report}{which type of report to make; diff is the difference between
the two variables by group}
}
\value{
matrix of results
}
\description{
\code{compare_binary} takes in a variable e.g. union
and runs bivariate regression of x on treatment (for summary statistics)
}
