% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{deg2rad}
\alias{deg2rad}
\title{Miscelaneous math functions}
\usage{
deg2rad(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
Convert angles from degrees to radians.
}

