% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{dropCovFromFormla}
\alias{dropCovFromFormla}
\title{Drop a Covariate from a Formula}
\usage{
dropCovFromFormla(covs, formla)
}
\arguments{
\item{covs}{should be a list of variable names}

\item{formla}{which formula to drop covariates from}
}
\value{
formula
}
\description{
\code{dropCovFromFormla} adds drops some covariates from a
formula; covs should be a list of variable names
}
\examples{
formla <- y ~ x + w + z
dropCovFromFormla(list("w","z"), formla)

dropCovFromFormla("z", formla)

}
