% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{dropCovFromFormla}
\alias{dropCovFromFormla}
\title{compareBinary}
\usage{
dropCovFromFormla(covs, formla)
}
\arguments{
\item{covs}{should be a list of variable names}

\item{formla}{which formula to drop covariates from}
}
\value{
formula
}
\description{
\code{compareBinary} drops some covariates from a formula;
  covs should be a list of variable names
}
