% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip vector names and values}
\usage{
flip(x)
}
\arguments{
\item{x}{A named vector.}
}
\value{
A named character vector.
}
\description{
Simply flips or switches values and names of named vectors. This is e.g.
useful when named character vectors are used a mapping.
}
\examples{
x <- structure(1:5, names = LETTERS[1:5])
x
flip(x)
flip(flip(x))


y <- structure(c("a", "b", "c"), names = c("abel", "bob", "charlie"))
y["bob"]
flip(y)["c"]
}
\author{
Anders Ellern Bilgrau
}

