% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip vector names and values}
\usage{
flip(x, convert = TRUE)
}
\arguments{
\item{x}{A named vector.}

\item{convert}{logical. If \code{TRUE}, an attempt will be made to coerce 
the flipped values into a \code{numeric}.}
}
\value{
A named character vector.
}
\description{
Simply flips or switches values and names of named vectors. This is e.g. 
useful when named character vectors are used a mapping.
}
\examples{
x <- structure(1:5, names = LETTERS[1:5])
x
flip(x)
flip(flip(x))


y <- structure(c("a", "b", "c", "d", "e", "f", "g"), 
   names = c("albert", "bob", "charlie", "dennis", "eddie", "frank", "hugo"))
y["hugo"]
flip(y)["a"]
}
\author{
Anders Ellern Bilgrau
}
