% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{get_group}
\alias{get_group}
\title{get_group}
\usage{
get_group(df, idname, tname, treatname)
}
\arguments{
\item{df}{the data.frame used in the function}

\item{idname}{name of column that holds the unit id}

\item{tname}{name of column that holds the time period}

\item{treatname}{name of column with the treatment indicator}
}
\description{
A function to calculate a unit's group in a panel data setting
 with a binary treatment and staggered treatment adoption and where
 there is a column in the data indicating whether or not a unit is treated
}
