% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{get_Yibar_pre}
\alias{get_Yibar_pre}
\title{get_Yibar_pre}
\usage{
get_Yibar_pre(df, idname, yname, tname, gname)
}
\arguments{
\item{df}{the data.frame used in the function}

\item{idname}{name of column that holds the unit id}

\item{yname}{name of column containing the outcome (or other variable)
for which to calculate its outcome in the immediate pre-treatment period}

\item{tname}{name of column that holds the time period}

\item{gname}{name of column containing the unit's group}
}
\description{
A function to calculate average outcomes for units in
 their pre-treatment periods (this function can also be used to recover
 pre-treatment averages of covariates, etc.).
 For units that do not
 participate in the treatment (and therefore have group==0), the
 function calculates their overall average outcome.
}
