% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{get_YiGmin1}
\alias{get_YiGmin1}
\title{get_YiGmin1}
\usage{
get_YiGmin1(df, idname, yname, tname, gname)
}
\arguments{
\item{df}{the data.frame used in the function}

\item{idname}{name of column that holds the unit id}

\item{yname}{name of column containing the outcome (or other variable) 
for which to calculate its outcome in the immediate pre-treatment period}

\item{tname}{name of column that holds the time period}

\item{gname}{name of column containing the unit's group}
}
\description{
A function to calculate outcomes for units in the period 
 right before they become treated (this function can also be used to recover 
 covariates, etc. in the period right before a unit becomes treated).
 For units that do not 
 participate in the treatment (and therefore have group==0), they are
 assigned their outcome in the last period.
}
