% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{get_Yit}
\alias{get_Yit}
\title{get_Yit}
\usage{
get_Yit(df, idname, yname, tname, gname)
}
\arguments{
\item{df}{the data.frame used in the function}

\item{idname}{name of column that holds the unit id}

\item{yname}{name of column containing the outcome (or other variable)
for which to calculate its outcome in the immediate pre-treatment period}

\item{tname}{name of column that holds the time period}

\item{gname}{name of column containing the unit's group}
}
\value{
a vector of outcomes in period t, the vector
 will have the length nT (i.e., this is returned for
 each element in the panel, not for a particular period)
}
\description{
A function to calculate outcomes for units in a particular
 time period `tp` in a panel data setting (this function can also
 be used to recover covariates, etc. in the first period).
}
