% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{ggMarginPlot}
\alias{ggMarginPlot}
\title{Margin Plot}
\usage{
ggMarginPlot(object, include.legend = TRUE, xlab = "Index",
  ylab = "Margin", main = "Margin Plot\\n", ...)
}
\description{
Margin Plot
}

