% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/math.R
\name{hypot}
\alias{hypot}
\title{Miscelaneous math functions}
\usage{
hypot(x, y)
}
\arguments{
\item{x}{A numeric value representing one leg of the triangle.}

\item{y}{A numeric value representing the other leg of the triangle.}
}
\value{
The length of the hypotenuse.
}
\description{
Calculate the length of the hypotenuse of a right-angled triangle.
}

