% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inrange.R
\name{inrange}
\alias{inrange}
\title{Within Range}
\usage{
inrange(x, lower, upper, strict = TRUE)
}
\arguments{
\item{x}{A number, or vector thereof.}

\item{lower}{Lower bound of range.}

\item{upper}{Upper bound of range.}

\item{strict}{Logical indicating whether or not to use strict inequalities
when checking bounds. Default is \code{TRUE}.}
}
\description{
Check whether a set of values are within a given range.
}
\examples{
inrange(3, 0, 1)
mean(inrange(rnorm(1e+06), -3, 3))
}

