% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{invertEcdf}
\alias{invertEcdf}
\title{Invert Ecdf}
\usage{
invertEcdf(df)
}
\arguments{
\item{df}{an ecdf object}
}
\value{
stepfun object that contains the quantiles of the df
}
\description{
take an ecdf object and invert it to get a step-quantile
 function
}
