% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strings.R
\name{is_digit}
\alias{is_digit}
\title{String Methods}
\usage{
is_digit(string)
}
\arguments{
\item{string}{An object that inherits from class \code{"character"}.}
}
\value{
Returns \code{TRUE} if all characters in the string are digits and
  there is at least one character, FALSE otherwise.
}
\description{
General test if characters in a string are digits.
}

