% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{lhs.vars}
\alias{lhs.vars}
\title{Left-hand Side Variables}
\usage{
lhs.vars(formla)
}
\arguments{
\item{formla}{a formula}
}
\value{
vector of variable names
}
\description{
Take a formula and return a vector of the variables
 on the left hand side, it will return NULL for a one sided formula
}
\examples{
ff <- yvar ~ x1 + x2
lhs.vars(ff)
}
