% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{loadPackages}
\alias{loadPackages}
\alias{unloadPackages}
\title{Load/Unload Multiple Packages}
\usage{
loadPackages(pkgs, ...)

unloadPackages(pkgs, ...)
}
\arguments{
\item{pkgs}{Character vector of package names.}

\item{...}{Additional optional arguments to be passed onto \code{library} or
\code{detach}.}
}
\description{
Load or unload multiple packages at once, rather than having to call
\code{library} or \code{detach} separately for each.
}

