% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{makeBalancedPanel}
\alias{makeBalancedPanel}
\title{makeBalancedPanel}
\usage{
makeBalancedPanel(data, idname, tname)
}
\arguments{
\item{data}{data.frame used in function}

\item{idname}{unique id}

\item{tname}{time period name}
}
\value{
data.frame that is a balanced panel
}
\description{
This function drops observations from data.frame
 that are not part of balanced panel data set.
}
