% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{makeBalancedPanel}
\alias{makeBalancedPanel}
\title{makeBalancedPanel}
\usage{
makeBalancedPanel(data, idname, tname, return_data.table = FALSE)
}
\arguments{
\item{data}{data.frame used in function}

\item{idname}{unique id}

\item{tname}{time period name}

\item{return_data.table}{if TRUE, make_balanced_panel will
return a data.table rather than a data.frame.  Default
is FALSE.}
}
\description{
Legacy version of `make_balanced_panel`,
 please use that function name going forward, though this will
 still work for now.
}
\keyword{internal}
