% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{makeDist}
\alias{makeDist}
\title{makeDist}
\usage{
makeDist(
  x,
  Fx,
  sorted = FALSE,
  rearrange = FALSE,
  force01 = FALSE,
  method = "constant"
)
}
\arguments{
\item{x}{vector of values}

\item{Fx}{vector of the distribution function values}

\item{sorted}{boolean indicating whether or not x is already sorted;
computation is somewhat faster if already sorted}

\item{rearrange}{boolean indicating whether or not should monotize
distribution function}

\item{force01}{boolean indicating whether or not to force the values of
the distribution function (i.e. Fx) to be between 0 and 1}

\item{method}{which method to pass to \code{approxfun} to approximate the
distribution function.  Default is "constant"; other possible choice is
"linear".  "constant" returns a step function, just like an empirical
cdf; "linear" linearly interpolates between neighboring points.}
}
\description{
Legacy name of `make_dist` function, please
 use that function instead.  This function will eventually
 be deleted.
}
\keyword{internal}
