% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mat}
\alias{mat}
\alias{mat.character}
\alias{mat.default}
\alias{mat.list}
\title{NumPy style matrix function}
\usage{
mat(x, ...)

\method{mat}{default}(..., rows = TRUE)

\method{mat}{character}(x)

\method{mat}{list}(x, rows = TRUE)
}
\description{
Returns a matrix from a character vector or list of vectors.
}
\examples{
## Using character vectors
mat('1, 2, 3; 4, 5, 6')  # similar to numpy.mat from Python extension NumPy
(m <- mat(paste('exp(', 1:9, ')')))
resize(m, 3, 3)
resize(m, 3, 3, byrow = FALSE)
matrix(exp(1:9), 3, 3)

## Using a list
mat(list(1:3, 4:6, 7:9))
}

