% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num.R, R/stats.R
\name{most_freq}
\alias{most_freq}
\title{Most Frequent Value}
\usage{
most_freq(x, na.rm = TRUE)

most_freq(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Logical indicating whether \code{NA} values should be stripped
before the computation prceeds.}

\item{x}{A vector of numbers or characters.}
}
\description{
\code{most_freq} returns the most frequent value in \code{x}.

Compute the mode (i.e., most frequent element).
}

