% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{multiplot}
\alias{multiplot}
\title{Multiple Plots}
\usage{
multiplot(..., cols = 1, layout = NULL)
}
\arguments{
\item{...}{Plots of class \code{"ggplot"}.}

\item{cols}{Desired number of coluns in final layout of plot.}

\item{layout}{Desired layout of plots.}
}
\description{
Multiple \code{"ggplot"} plots on a single page.
}

