% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hybrid.R
\name{addNodeFactor}
\alias{addNodeFactor}
\title{Create CART-like Dummy Variables}
\usage{
addNodeFactor(object, newdata, important = FALSE, ...)
}
\arguments{
\item{object}{An object that inherits from class \code{rpart}.}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{important}{Logical indicating whether to include only important
variables. Default is \code{FALSE}.}

\item{...}{Additional optional arguments. At present, no optional arguments
  are used.}
}
\value{
A data frame containing the original data with an additional factor
  variable containing terminal node assignments.
}
\description{
Augments the supplied data with an additional factor variable describing
terminal node assignment.
}
\examples{
#
# Boston housing data
#

# Load packages and data
library(rpart)  # for CART-like decision trees
library(earth)  # for MARS-like regression models
data(Boston, package = "MASS")  # Boston housing data

# CART model
boston_cart <- rpart(medv ~ ., data = Boston, cp = 0.005)
Boston2 <- addNodeFactor(boston_cart)

# MARS model
earth(medv ~ ., data = Boston, degree = 2, linpreds = TRUE)

# Hybrid model
earth(medv ~ ., data = Boston2, degree = 2, linpreds = TRUE)
}

