% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{nunique}
\alias{nunique}
\alias{nunique.default}
\alias{nunique.matrix}
\alias{nunique.data.frame}
\title{Number of unique elements}
\usage{
nunique(x)

\method{nunique}{default}(x)

\method{nunique}{matrix}(x)

\method{nunique}{data.frame}(x)
}
\arguments{
\item{x}{A vector, matrix, or data frame.}
}
\description{
Return the number of unique values which its argument has.
}
\examples{
nunique(c(1, 1, 2))
nunique(iris$Species)
}
