% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nunique.R
\name{nunique}
\alias{nunique}
\alias{nunique.data.frame}
\alias{nunique.default}
\alias{nunique.matrix}
\title{The Number of Unique Elements}
\usage{
nunique(x)

\method{nunique}{default}(x)

\method{nunique}{matrix}(x)

\method{nunique}{data.frame}(x)
}
\arguments{
\item{x}{A vector, matrix, or data frame.}
}
\description{
Return the number of unique values which its argument has.
}

