% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ones.R
\name{ones}
\alias{ones}
\alias{zeros}
\title{Create an array of ones or zeros}
\usage{
ones(...)

zeros(...)
}
\arguments{
\item{\dots}{A vector or list of integers giving the dimension of the array.
A dimnames argument can be supplied.}
}
\value{
Returns an \code{array}, \code{matrix}, or \code{vector} of the
  dimension specified by the arguments.
}
\description{
Easy interface for creating an array of supplied dimension of ones and zeros.
Simply a wrapper for \code{array}. Built for convinienece and not speed.
}
\examples{
ones(6)
ones(3,5)
ones(c(2,3,3))
ones(3,3,3, dimnames=list(LETTERS[1:3], letters[1:3], LETTERS[4:6]))
zeros(4, 3)
zeros(list(5, 3, 2))
zeros(c(3, 4), dimnames=list(LETTERS[1:3], letters[9:12]))
}
\seealso{
\code{\link{array}}
}

