% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%notin\%}
\alias{\%notin\%}
\alias{\%+=\%}
\alias{\%+=\%.default}
\alias{\%-=\%}
\alias{\%-=\%.default}
\alias{\%*=\%}
\alias{\%*=\%.default}
\alias{\%/=\%}
\alias{\%/=\%.default}
\alias{\%+\%}
\alias{\%+\%.default}
\alias{\%+\%.character}
\alias{\%-\%}
\alias{\%-\%.character}
\title{Additional operators}
\usage{
x \%notin\% y

x \%+=\% y

\method{\%+=\%}{default}(x, y)

x \%-=\% y

\method{\%-=\%}{default}(x, y)

x \%*=\% y

\method{\%*=\%}{default}(x, y)

x \%/=\% y

\method{\%/=\%}{default}(x, y)

x \%+\% y

\method{\%+\%}{default}(x, y)

\method{\%+\%}{character}(x, y)

x \%-\% y

\method{\%-\%}{character}(x, y)
}
\arguments{
\item{x}{The values to be matched.}

\item{y}{The values to be matched against.}
}
\description{
Returns a logical vector indicating if there is not a match for its left 
operand.
}
