% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pcor}
\alias{pcor}
\title{Compute first order partial correlations}
\usage{
pcor(X, k)
}
\arguments{
\item{X}{an numeric correlation matrix}

\item{k}{an integer. The variable to condition on.}
}
\value{
A matrix of the same size as \code{X} with \code{NaN}s in the
  \code{k} row and column.
}
\description{
This function computes the first order partial correlation given the
\code{k}'th variable.
}
\examples{
mat <- replicate(4, rnorm(10))
X <- cor(mat)
pcor(X, 2)
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\link{\code{cor}}.
}

