% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/battery.R
\name{plot.battery_target}
\alias{plot.battery_target}
\title{Plot the Output of a Battery Target Run}
\usage{
\method{plot}{battery_target}(x, ...)
}
\arguments{
\item{x}{An object that inherits from class \code{"battery_target"}.}

\item{...}{Additional optional arguments to be passed on to \code{heatmap}.}
}
\description{
This visualizes the output of a battery target run using a heatmap of the
variable importance matrix.
}

