% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotHierarchicalEdgeBundles}
\alias{plotHierarchicalEdgeBundles}
\title{Hierarchical Edge Bundling}
\usage{
plotHierarchicalEdgeBundles(phylo, graph, beta = 0.5, use.only.mrca = FALSE,
  simplify = FALSE, ..., args.lines = list(), debug = FALSE)
}
\arguments{
\item{phylo}{A `phylo` object.}

\item{graph}{A \code{igraph} object.}

\item{beta}{The amount of bundling.}

\item{use.only.mrca}{Should the only the most recent common ancestor or the
full shortest path be used?}

\item{simplify}{Simplify the paths by taking the convex hull.}

\item{...}{Arguments passed to \code{\link[ape]{plot.phylo}}.}

\item{args.lines}{A list of arguments passed to \code{lines}.}

\item{debug}{Plot some extra info.}
}
\value{
Plots to a new device.
}
\description{
Visualization of networks using Hierarchical Edge Bundles.
Currently, Bezier curves are used and not B-splines for the bundling.
Hopefully, this will be updated in future versions.
}
\examples{
if (require("igraph")) {
  n <- 25
  soft <- function(a, c) sign(a)*pmax(abs(a) - c, 0) # Soft thresholding

  # Create graph
  adj <- abs(cor(matrix(rnorm(n^2), n, n)))
  graph <- graph.adjacency(adj, mode = "un", weighted = TRUE, diag = FALSE)
  V(graph)$name <- apply(combn(LETTERS, 2), 2, paste0, collapse = "")[1:n]
  E(graph)$weight <- E(graph)$weight^2
  E(graph)$color <- alp("black", E(graph)$weight/max(E(graph)$weight))

  wt <- fastgreedy.community(graph)
  phylo <- asPhylo(wt)

  par(mfrow = c(1,3))
  plot(graph, layout = layout.circle)
  plot(phylo, type = "fan")
  plotHierarchicalEdgeBundles(phylo, graph, type = "fan", beta = 0.90,
                              debug = FALSE,
                              simplify = TRUE,
                              args.lines = list(col = alp("tomato", 30)))
}
}
\author{
Anders Ellern Bilgrau
}
\references{
Holten, Danny. "Hierarchical edge bundles: Visualization of adjacency
  relations in hierarchical data." Visualization and Computer Graphics, IEEE
  Transactions on 12.5 (2006): 741-748.
}
\seealso{
See \code{\link[ape]{plot.phylo}}.
}

