% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotHierarchicalEdgeBundles}
\alias{plotHierarchicalEdgeBundles}
\title{Hierarchical Edge Bundling}
\usage{
plotHierarchicalEdgeBundles(phylo, graph, beta = 0.5, use.only.mrca = TRUE,
  simplify = FALSE, ..., args.lines = list(), debug = FALSE)
}
\arguments{
\item{phylo}{A `phylo` object.}

\item{graph}{A \code{igraph} object.}

\item{beta}{The amount of bundling.}

\item{use.only.mrca}{Should the only the most recent common ancestor or the
full shortest path be used?}

\item{simplify}{Simplify the paths by taking the convex hull.}

\item{...}{Arguments passed to \code{\link[ape]{plot.phylo}}.}

\item{args.lines}{A list of arguments passed to \code{lines}.}

\item{debug}{Plot some extra info.}
}
\value{
Plots to a new device.
}
\description{
Visualization of networks using Hierarchical Edge Bundles.
Currently, Bezier curves are used and not B-splines for the bundling.
Hopefully, this will be updated in future versions.
}
\examples{
if (require("igraph")) {
  n <- 100
  graph <- erdos.renyi.game(n = n, p = 0.02)
  V(graph)$name <- apply(combn(LETTERS, 2), 2, paste0, collapse = "")[1:n]
  wt <- walktrap.community(graph, modularity=TRUE)
  phylo <- asPhylo(wt)
  plotHierarchicalEdgeBundles(phylo,
                              graph,
                              beta = 0.2,
                              use.only.mrca = FALSE,
                              type = "fan", #or "unrooted"
                              use.edge.length = TRUE,
                              simplify = FALSE,
                              lab4ut = "axial",
                              args.lines = list(col = "#FF00FFFF"))
}
}
\author{
Anders Ellern Bilgrau
}
\references{
Holten, Danny. "Hierarchical edge bundles: Visualization of adjacency
  relations in hierarchical data." Visualization and Computer Graphics, IEEE
  Transactions on 12.5 (2006): 741-748.
}
\seealso{
See \code{\link[ape]{plot.phylo}}.
}

