% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotHierarchicalEdgeBundles}
\alias{plotHierarchicalEdgeBundles}
\title{Hierarchical Edge Bundling}
\usage{
plotHierarchicalEdgeBundles(phylo, graph, beta = 0.5, include.mrca = FALSE,
  simplify = FALSE, ..., args.lines = list(), debug = FALSE, v.use.only,
  e.use.only)
}
\arguments{
\item{phylo}{A `phylo` object.}

\item{graph}{A \code{igraph} object.}

\item{beta}{The amount of bundling.}

\item{include.mrca}{Should the only the most recent common ancestor be used
in the shortest path used for the splines?}

\item{simplify}{Simplify the paths by taking the convex hull of the control
points. Can sometimes yield better results.}

\item{...}{Arguments passed to \code{\link[ape]{plot.phylo}}.}

\item{args.lines}{A list of arguments passed to \code{lines}.}

\item{debug}{Plot some extra debug info.}

\item{v.use.only}{An integer vector giving the nodes from which edges are
to be drawn. E.g. \code{use.only = 1} will only plot the edges from
vertex 1.}

\item{e.use.only}{An integer vector giving the edges to be drawn.
E.g. \code{use.only = 1} will only draw the first edge.}
}
\value{
Plots to a new device.
}
\description{
Visualization of networks using hierarchical edge bundles. Plot a graph
using a hierchical tree as guies for the edges.
}
\examples{
library("igraph")
library("ape")
n <- 25

# Create graph
adj <- abs(cor(matrix(rnorm(n^2), n, n)))
rownames(adj) <- colnames(adj) <-
  apply(combn(LETTERS, 2), 2, paste0, collapse = "")[1:n]
graph <- graph.adjacency(adj, mode = "un", weighted = TRUE, diag = FALSE)
E(graph)$weight <- E(graph)$weight^2
E(graph)$color <- alp("black", E(graph)$weight/max(E(graph)$weight))

phylo <- as.phylo(hclust(as.dist(1 - adj), method = "ward.D"))

par(mfrow = c(1,3))
plot(phylo, type = "fan")
plotHierarchicalEdgeBundles(phylo, graph, type = "fan", beta = 0.95,
                            args.lines = list(col = alp("steelblue", 90)))
plotHierarchicalEdgeBundles(phylo, graph, type = "fan", beta = 0.85,
                            args.lines = list(col = alp("steelblue", 90)))

# Extra control of plotting and debugging
par(mfrow = c(1,2))
plot(phylo, type = "fan")
plotHierarchicalEdgeBundles(phylo, graph, type = "fan", beta = 0.95,
                            v.use.only = 1, debug = FALSE,
                            args.lines = list(col = alp("red", 1), lwd = 2))
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\references{
Holten, Danny. "Hierarchical edge bundles: Visualization of adjacency
  relations in hierarchical data." Visualization and Computer Graphics, IEEE
  Transactions on 12.5 (2006): 741-748.
}
\seealso{
See \code{\link[ape]{plot.phylo}}.
}

