% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotModuleGraph.R
\name{plotModuleGraph}
\alias{plotModuleGraph}
\title{Plot the graphical representation of a weighted adjacency amatrix}
\usage{
plotModuleGraph(amat, diff.exprs = NULL, labels = rownames(amat),
  vcol = "Black", ecol = "Black", layout = layout.circle, ...)
}
\arguments{
\item{amat}{A numeric adjacency a matrix with values between 0 and 1.}

\item{diff.exprs}{A numeric vector of lencth \code{ncol(amat)} giving the
importance (e.g. evidence for differential expression) of the node/gene.}

\item{labels}{The labels of the nodes.}

\item{vcol}{The colors of the nodes. Default is black.}

\item{ecol}{The colour of the edges. Default is black.}

\item{layout}{A function providing the layout of the graph.
See \link[igraph]{layout}.}

\item{\dots}{Aguments passed to \code{plot.igraph}.}
}
\value{
Invisibly returns the igraph-object and the layout in a \code{list}.
}
\description{
A function of for plotting the graphical representation of a weigted
weighted adjacencey amatrix. Build for convenience and not for speed.
}
\examples{
# Construct a random adjacency matrix
amat <- replicate(10, runif(10))
amat[upper.tri(amat)] <- amat[lower.tri(amat)]

# Plot the graph
plotModuleGraph(amat,
                labels = "",
                vcol = 1:10,
                diff.exprs = 3 + 3*runif(10))
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}

