% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{rad2deg}
\alias{rad2deg}
\title{Miscelaneous math functions}
\usage{
rad2deg(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
Convert angles from radians to degrees.
}

