% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resave.R
\name{resave}
\alias{resave}
\title{Update or append objects to \code{.RData} file}
\usage{
resave(..., list = character(), file)
}
\arguments{
\item{list}{A character of the names to be resaved.}

\item{file}{A character of length 1 giving the path and \code{.RData} file
to resave the objects into.}

\item{\dots}{The objects to be saved, appended or overwritten.}
}
\value{
Returns nothing. If \code{.RData}-file specified by \code{file}
  does not exist \code{resave} functions exactly as \code{save} and creates
  a new file. If it does exist the \code{file} is overwritten with the
  updated objects.
}
\description{
Function for appending or updating objects to an existing \code{.RData} file.
\code{resave} updates existing objects in an .RData file. If the
resaved objects do not exist in the file, the objects are added to the
file. If the file does not exist, \code{resave} functions as \code{save}.
}
\note{
This is a modified version of a function created by user "flodel" at
  \href{stack overflow.}{http://stackoverflow.com/questions/11813096/updating-an-existing-rdata-file}
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
file <- file.path(tempdir(), "myFile.RData") # File
a <- list(LETTERS)

resave(a, file = file)  # Save object "a"

b <- pi  # Create a new object "b"
a <- c(a, list(letters))  # Modify "a"

resave(a, b, file = file)  # Update a and add b
}
\seealso{
See Also: \code{\link{save}}
}

