% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{rhs.vars}
\alias{rhs.vars}
\title{rhs.vars}
\usage{
rhs.vars(formla)
}
\arguments{
\item{formla}{a formula}
}
\description{
Legacy version of `rhs_vars`, please use that
 function instead.  This function will eventually be deleted.
}
\keyword{internal}
