% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{set_seed}
\alias{set_seed}
\title{Specify Seeds}
\usage{
set_seed(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
Returns NULL, invisibly.
}
\description{
Specify seeds using character vector.
}
\examples{
set_seed("Some random numbers")
rnorm(3)  # should give: 1.7500983 -0.1093635 -0.9958618
}

