% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{str2int}
\alias{str2int}
\title{Convert Characters to Integers}
\usage{
str2int(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
An integer (not necessarily unique).
}
\description{
Convert one-dimensional character objects to integers.
}
\examples{
# Should all produce the same integer (currently results in 124)
str2int("Hello World")
str2int("hello world")
str2int("hello   world")
str2int("hello.world")
str2int("hello...world")
str2int("hello-world")
str2int("hello - world")
str2int("hello---world")
str2int("helloworld")
str2int("HelloWorld")
str2int("H._.e L. - lo -.  W__or l- .. d")

set.seed(str2int("Some random numbers"))
rnorm(3)  # should give: 1.7500983 -0.1093635 -0.9958618
set.seed(str2int("Some more random numbers"))
rnorm(3)  # should give: 0.007765185 -1.138536203  0.091017129

# Potential issues
str2int("abc")  # does not provide unique integers
str2int("cba")
}

