% Generated by roxygen2 (4.0.2): do not edit by hand
\name{straightenedBezier}
\alias{straightenedBSpline}
\alias{straightenedBezier}
\title{Straightened Spline curves}
\usage{
straightenedBezier(x, y, beta = 0.5, n.evals = 100)

straightenedBSpline(x, y, beta = 0.5, order = 4, n.evals = 100)
}
\arguments{
\item{x}{A numeric vector giving the x-coordinates of the control points.
Alternatively, it can be a \code{list} (or \code{data.frame}) containing
two vectors of control points.}

\item{y}{A numeric vector giving the y-coordinates of the control points.}

\item{beta}{A numeric values that control amount of straightening.
A \code{beta} of \eqn{1} is a normal Bezier curve and where as a value of
\eqn{0} is a straight line.}

\item{n.evals}{The number of evaluations of the Bezier curve.}

\item{order}{The order of the B-spline. See \code{\link{bSpline}}.}
}
\value{
Returns a matrix of spline evaluations.
}
\description{
Function for straightning Bezier or B-spline curves.
}
\examples{
x <- list()
}
\author{
Anders Ellern Bilgrau
}
\keyword{internal}

