% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMisc.R
\name{time_invariant_to_panel}
\alias{time_invariant_to_panel}
\title{time_invariant_to_panel}
\usage{
time_invariant_to_panel(x, df, idname, balanced_panel = FALSE)
}
\arguments{
\item{x}{a vector of length equal to the number of unique ids in df.}

\item{df}{the data.frame used in the function}

\item{idname}{name of column that holds the unit id}

\item{balanced_panel}{a logical indicating whether the panel is balanced.}
}
\value{
a vector of length equal to the number of rows in df.
}
\description{
This function takes a time-invariant variable and repeats it
 for each period in a panel data set.
}
