% Generated by roxygen2 (4.0.1): do not edit by hand
\name{without}
\alias{\%w/o\%}
\alias{without}
\title{The relative complement or set difference}
\usage{
without(x, y)
x \%w/o\% y
}
\arguments{
\item{x}{A character or numeric vector.}

\item{y}{A character or numeric vector.}
}
\value{
The functions returns a vector of the elements in x which does not appear in y.
}
\description{
The relative complement of x with respect to y, i.e. x\\y or x minus y.
}
\note{
These functions are identical to \code{\link{setdiff}} but sometimes more easy to use.
}
\examples{
print(x <- LETTERS[1:10])
print(y <- LETTERS[5:15])

x \%w/o\% y
without(x, y)
setdiff(x, y)
}

