% Generated by roxygen2 (4.0.2): do not edit by hand
\name{without}
\alias{\%\\\%}
\alias{\%w/o\%}
\alias{without}
\title{The relative complement or set difference}
\usage{
without(x, y)
x \%w/o\% y
}
\arguments{
\item{x}{A character or numeric vector.}

\item{y}{A character or numeric vector.}
}
\value{
The functions returns a vector of the elements in x which does not appear in y.
}
\description{
The page decribes different interfaces for the relative complement of x with
respect to y, i.e. x\\y or x minus y. It is faster than
}
\note{
These functions are identical to \code{\link{setdiff}} but sometimes more easy to use.
}
\examples{
print(A <- LETTERS[1:10])
print(B <- LETTERS[5:15])

A \%w/o\% B
#A \%\\\% B     # More mathematical notation, single backslash only!
without(A, B)
setdiff(A, B)

\dontrun{
library(microbenchmark)
microbenchmark(without(A, B),
               setdiff(A, B))
}
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

